+++ test/dummy/app/assets/bundles/application.js
@@ -1,2 +1,2 @@
-!function(e){function t(t){for(var r,a,c=t[0],u=t[1],l=t[2],s=0,p=[];s<c.length;s++)a=c[s],o[a]&&p.push(o[a][0]),o[a]=0;for(r in u)Object.prototype.hasOwnProperty.call(u,r)&&(e[r]=u[r]);for(f&&f(t);p.length;)p.shift()();return i.push.apply(i,l||[]),n()}function n(){for(var e,t=0;t<i.length;t++){for(var n=i[t],r=!0,c=1;c<n.length;c++){var u=n[c];0!==o[u]&&(r=!1)}r&&(i.splice(t--,1),e=a(a.s=n[0]))}return e}var r={},o={0:0},i=[];function a(t){if(r[t])return r[t].exports;var n=r[t]={i:t,l:!1,exports:{}};return e[t].call(n.exports,n,n.exports,a),n.l=!0,n.exports}a.m=e,a.c=r,a.d=function(e,t,n){a.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},a.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},a.t=function(e,t){if(1&t&&(e=a(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(a.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var r in e)a.d(n,r,function(t){return e[t]}.bind(null,r));return n},a.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return a.d(t,"a",t),t},a.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},a.p="";var c=window.webpackJsonp=window.webpackJsonp||[],u=c.push.bind(c);c.push=t,c=c.slice();for(var l=0;l<c.length;l++)t(c[l]);var f=u;i.push([143,1]),n()}({101:function(e,t,n){},102:function(e,t,n){},143:function(e,t,n){"use strict";n.r(t);n(101),n(102);var r=n(98);n.n(r).a.start();var o=n(1),i=n(99);n(4),n(5),n(6),n(3),n(34),n(10),n(11),n(2),n(7),n(8);function a(e){return(a="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function c(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function u(e,t){return!t||"object"!==a(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function l(e){return(l=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function f(e,t){return(f=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var s=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=u(this,l(t).call(this,n))).notice=n.notice,e.alert=n.alert,e.warning=n.warning,e.hide=n.hide,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&f(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"setNotice",value:function(e){this.notice=e}},{key:"setAlert",value:function(e){this.alert=e}},{key:"setWarning",value:function(e){this.warning=e}},{key:"render",value:function(){var e=document.querySelector(".flash");e.classList.remove("notice"),e.classList.remove("alert"),e.classList.remove("warning"),null!=this.notice?(e.classList.add("notice"),document.querySelector(".flash p").textContent=this.notice):null!=this.alert?(e.classList.add("alert"),document.querySelector(".flash p").textContent=this.alert):null!=this.warning&&(e.classList.add("warning"),document.querySelector(".flash p").textContent=this.warning),e.classList.remove("none"),this.hide&&this.hideAfterTime()}},{key:"hideAfterTime",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:4e3;setTimeout(function(){document.querySelector(".flash").classList.add("none")},e)}}])&&c(n.prototype,r),i&&c(n,i),t}(),p={disconnectedForTooLong:function(){new s({alert:"You have been disconnected from the server for too long. Reload page!",hide:!1}).render()}},y=(n(30),n(0)),m=n.n(y),d=n(16),b="ADD_ARTICLES",h="SET_ARTICLES",v="UPDATE_ARTICLE",w="ADD_COMMENTS",g="REMOVE_COMMENT",O="SET_COMMENTS";function _(e){return{type:b,payload:{articles:e}}}function j(e){return{type:h,payload:{articles:e}}}function S(e,t){return{type:v,payload:{article:e,index:t}}}function E(e,t){return{type:O,payload:{comments:e,articleId:t}}}var P=n(20),k=(n(19),n(41),n(31),n(42),n(25),"REMOVE_ARTICLE");function A(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}n(70),n(55),n(56),n(57);function I(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function C(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){T(e,t,n[t])})}return e}function T(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var x=Object(P.a)({articles:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case b:return[].concat(A(e),A(t.payload.articles));case k:return e.filter(function(e){return e.id!==t.payload.id});case h:return A(t.payload.articles);case v:return[].concat(A(e.slice(0,t.payload.index)),[t.payload.article],A(e.slice(t.payload.index+1)));default:return e}},comments:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=arguments.length>1?arguments[1]:void 0;switch(t.type){case w:return C({},e,T({},t.payload.articleId,[].concat(I(e[t.payload.articleId]),I(t.payload.comments))));case g:return null==e[t.payload.articleId]?e:C({},e,T({},t.payload.articleId,e[t.payload.articleId].filter(function(e){return e.id!==t.payload.id})));case O:return T({},t.payload.articleId,t.payload.comments);case"UPDATE_COMMENT":var n=t.payload.articleId,r=t.payload.index;if(!r){var o=e[n].find(function(e){return e.id===t.payload.comment.id});r=e[n].indexOf(o)}return C({},e,T({},n,[].concat(I(e[n].slice(0,r)),[t.payload.comment],I(e[n].slice(r+1)))));default:return e}}}),R=Object(P.b)(x,{articles:[],comments:{}});function B(e){return(B="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function z(e,t){return!t||"object"!==B(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function D(e){return(D=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function N(e,t){return(N=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var M=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),z(this,D(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&N(e,t)}(t,o["Views"].Base),t}();n(94);function L(e){return(L="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function q(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function V(e,t){return!t||"object"!==L(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function U(e){return(U=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function F(e,t){return(F=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var W=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=V(this,U(t).call(this,n))).article=null,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&F(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"renderArticle",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null;null==this.article&&(this.connectWith(e,{receiver:"articleReceivedSignal"}),this.article=e),document.getElementById("article_title").textContent=this.article.title,document.getElementById("article_text").textContent=this.article.content;var t=document.getElementById("publish_article");t&&null!=this.article.publishedAt?t.style.display="none":t&&(t.style.display=""),this._handlePublishing(),this._updateEditLink()}},{key:"articleReceivedSignal",value:function(e){var t=this;switch(e){case"updated":this.article.reload().then(function(){t.article.applyChanges(),t.renderArticle()});break;case"destroyed":window.location.href="/user/articles?message=deleted"}}},{key:"_handlePublishing",value:function(){var e=this;document.getElementById("publish_article").addEventListener("click",function(t){t.preventDefault(),t.target.textContent="Publishing...",e.article.put("publish").then(function(){return document.getElementById("publish_article").outerHTML="<span>Published!</span>"}).catch(function(){t.target.textContent="Publish",new s({alert:"Connection error!"}).render()})})}},{key:"_updateEditLink",value:function(){var e=document.getElementById("edit_link").getAttribute("href");document.getElementById("edit_link").setAttribute("href",e.replace("/0/","/".concat(this.article.id,"/")))}}])&&q(n.prototype,r),i&&q(n,i),t}();n(129),n(131);function H(e){return(H="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function J(e,t){return!t||"object"!==H(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Y(e){return(Y=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function $(e,t){return($=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function G(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var K=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),J(this,Y(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&$(e,t)}(t,o["Models"].Base),t}();G(K,"identity","Article.Comment"),G(K,"remoteName","Comment"),G(K,"resources",{url:"/user/articles/:articleId/comments",paginate:{per:10},main:{url:"/articles/:articleId/comments",paginate:{per:5,param:"page-num"}},admin:{url:"/admin/articles/:articleId/comments",paginate:{per:5}}}),G(K,"attributes",{author:{validations:{presence:!0}},text:{validations:{presence:!0,vulgarity:!0}},articleId:{type:"Int",validations:{presence:!0},remoteName:"article_id"},createdAt:{type:"Date",remoteName:"created_at"},updatedAt:{type:"Date",remoteName:"updated_at"},emotion:{type:"Int"},pinned:{type:"Boolean"},adminRate:{type:"Int",remoteName:"admin_rate"},approved:{type:"Boolean"}});var Q=K,X=(n(32),n(9)),Z=n.n(X),ee=function(e,t){return e.comments[t]};function te(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){ne(e,t,n[t])})}return e}function ne(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function re(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function oe(e){var t,n=e.comment,r=e.isAdmin,i=void 0!==r&&r,a=re(Object(y.useState)(!1),2),c=a[0],u=a[1],l=new o.Services.Date(n.createdAt).strftime("%d %b %y"),f=function(e){e.preventDefault(),u(!0),n.approved=!0,n.updateAttribute("approved").then(function(e){e.ok&&R.dispatch({type:"UPDATE_COMMENT",payload:{articleId:n.articleId,comment:new Q(te({},n,{approved:!0}))}})})};return m.a.createElement("p",{id:"comment_".concat(n.id)},m.a.createElement("b",null,n.author)," on ",m.a.createElement("i",null,l),i?(t=n.approved?m.a.createElement(m.a.Fragment,null,m.a.createElement("span",null,"approved")," |"," "):c?m.a.createElement(m.a.Fragment,null,m.a.createElement("span",null,"approving...")," |"," "):m.a.createElement(m.a.Fragment,null,m.a.createElement("a",{href:"/user/articles/".concat(n.articleId,"/comments/").concat(n.id,"/approve"),className:"approve",onClick:f},"approve")," ","|"," "),m.a.createElement(m.a.Fragment,null," ","(",t,m.a.createElement("a",{href:"/user/articles/".concat(n.articleId,"/comments/").concat(n.id,"/edit")},"edit")," ","|"," ",m.a.createElement("a",{href:"/user/articles/".concat(n.articleId,"/comments/").concat(n.id),"data-method":"delete","data-confirm":"Are you sure?"},"delete"),")")):"",m.a.createElement("br",null),n.text)}oe.propTypes={comment:Z.a.instanceOf(Q).isRequired,isAdmin:Z.a.bool};var ie=oe;function ae(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function ce(e){var t=e.articleId,n=e.isAdmin||!1,r=ae(Object(y.useState)(e.comments),2),o=r[0],i=r[1];if(Object(y.useEffect)(function(){var e=R.subscribe(function(){return i(ee(R.getState(),t))});return function(){e()}}),0===o.length)return m.a.createElement("p",null,"No comments.");var a=o.map(function(e){return m.a.createElement(ie,{key:e.id,comment:e,isAdmin:n})});return m.a.createElement(m.a.Fragment,null,a)}ce.propTypes={articleId:Z.a.number.isRequired,isAdmin:Z.a.bool,comments:Z.a.arrayOf(Z.a.instanceOf(Q)).isRequired};var ue=ce;function le(e){return(le="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function fe(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function se(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function pe(e,t){return!t||"object"!==le(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function ye(e){return(ye=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function me(e,t){return(me=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var de=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=pe(this,ye(t).call(this,n))).article=null,e.form=null,e.changes=null,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&me(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(e){R.dispatch(_([e])),this.article=e,this.connectWith(this.article),this._handleApplyingChanges(),this.form=new o.UI.Form({for:this.article}),this.form.render()}},{key:"renderComments",value:function(e){Q.all({articleId:e}).then(function(t){R.dispatch(E(t.resources,e)),Object(d.render)(m.a.createElement(ue,{articleId:e,comments:t.resources,isAdmin:!0}),document.getElementById("comments"))})}},{key:"receivedSignal",value:function(e,t){var n=this;switch(e){case"updating":if(document.querySelector("h1").getAttribute("data-mark")!==t.mark)new s({warning:"Uuups someone else started editing this article."}).render();break;case"updated":this.article.reload().then(function(){n.changes=n.article.changes(),n._displayChanges()});break;case"destroyed":window.location.href="/user/articles?message=deleted"}}},{key:"_displayChanges",value:function(){for(var e=0,t=Object.entries(this.changes);e<t.length;e++){var n=fe(t[e],1)[0],r=document.querySelector("a.apply_changes[data-for=".concat(this.article.getAttrRemoteName(n),"]"));r&&r.classList.remove("none")}}},{key:"_handleApplyingChanges",value:function(){for(var e=this,t=0,n=Array.from(document.querySelectorAll("a.apply_changes"));t<n.length;t++){n[t].addEventListener("click",function(t){t.preventDefault();var n=e.article.getAttrName(t.target.getAttribute("data-for"));e.article[n]=e.changes[n].is,e.form.fill(n),t.target.classList.add("none")})}}}])&&se(n.prototype,r),i&&se(n,i),t}();function be(e){return(be="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function he(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function ve(e,t){return!t||"object"!==be(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function we(e){return(we=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function ge(e,t){return(ge=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function Oe(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var _e=function(e){function t(e){var n;return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(n=ve(this,we(t).call(this,e))).published=null!=n.publishedAt,n}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&ge(e,t)}(t,o["Models"].Base),n=t,(r=[{key:"vulgarityLevel",value:function(){(null!=this.title&&/fuck/i.exec(this.title)||null!=this.content&&/fuck/i.exec(this.content))&&this.addErrorMessage("Article contains strong language.",{for:"base"})}},{key:"setDefaultValuesForAdminReview",value:function(){this.adminRate=null==this.adminRate?3:this.adminRate,this.categoryId=null==this.categoryId?6:this.categoryId,this.adminReviewStartedAt=Date.now()}}])&&he(n.prototype,r),i&&he(n,i),t}();Oe(_e,"identity","Article"),Oe(_e,"resources",{url:"/user/articles",paginate:{per:5},main:{url:"/articles",paginate:{per:3}},admin:{url:"/admin/articles",paginate:{per:4}}}),Oe(_e,"attributes",{title:{validations:{presence:!0,length:{within:[3,255]}}},content:{validations:{presence:!0,length:{minimum:100}},remoteName:"text"},createdAt:{type:"Date",remoteName:"created_at"},updatedAt:{type:"Date",remoteName:"updated_at"},commentsCount:{type:"Int",remoteName:"comments_count"},publishedAt:{type:"Date",remoteName:"published_at"},published:{},adminReview:{remoteName:"admin_review"},adminRate:{type:"Int",remoteName:"admin_rate"},categoryId:{type:"Int",remoteName:"category_id"},adminReviewStartedAt:{remoteName:"admin_review_started_at"}}),Oe(_e,"validate",["vulgarityLevel"]);var je=_e;function Se(e){var t=e.article,n=e.onArticleDestroyed;return m.a.createElement("tr",{id:"article_".concat(t.id)},m.a.createElement("td",null,t.title),m.a.createElement("td",null,t.content),m.a.createElement("td",{className:"comments_quantity"},t.commentsCount),m.a.createElement("td",{className:"published"},t.publishedAt?"yes":"no"),m.a.createElement("td",null,m.a.createElement("a",{href:"/user/articles/".concat(t.id)},"Show")," |",m.a.createElement("a",{href:"/user/articles/".concat(t.id,"/edit")},"Edit")," |",m.a.createElement("a",{href:"/user/articles/".concat(t.id),className:"delete_article",onClick:function(e){e.preventDefault(),confirm("Are you sure?")&&t.delete(null).then(function(e){var t;e.success&&R.dispatch((t=e.id,{type:k,payload:{id:t}})),n(e)})}},"Delete")))}Se.propTypes={article:Z.a.instanceOf(je).isRequired,onArticleDestroyed:Z.a.func.isRequired};var Ee=Se;function Pe(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function ke(e){var t=Pe(Object(y.useState)(e.articles),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=R.subscribe(function(){return r(R.getState().articles)});return function(){e()}});var o=n.map(function(t){return m.a.createElement(Ee,{key:t.id,article:t,onArticleDestroyed:e.onArticleDestroyed})});return m.a.createElement(m.a.Fragment,null,o)}ke.propTypes={articles:Z.a.arrayOf(Z.a.instanceOf(je)).isRequired,onArticleDestroyed:Z.a.func.isRequired};var Ae=ke;function Ie(e){return(Ie="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Ce(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Te(e,t){return!t||"object"!==Ie(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function xe(e){return(xe=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Re(e,t){return(Re=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Be=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Te(this,xe(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Re(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){this.layout=new M}},{key:"onArticleDestroyed",value:function(e){var t=new s;e.success?t.setNotice(e.notice):t.setAlert(e.alert),t.render()}},{key:"index",value:function(){var e=this;"deleted"===this.params.message&&new s({alert:"Article has been deleted."}).render();je.get("all").then(function(t){R.dispatch(j(t.resources)),Object(d.render)(m.a.createElement(Ae,{articles:t.resources,onArticleDestroyed:e.onArticleDestroyed}),document.getElementById("article_list"))})}},{key:"show",value:function(){var e=this;this.showView=new W,je.find(this.params.id).then(function(t){R.dispatch(j([t])),e.showView.renderArticle(t)}),Q.all({articleId:this.params.id}).then(function(t){R.dispatch(E(t.resources,e.params.id)),Object(d.render)(m.a.createElement(ue,{articleId:e.params.id,comments:t.resources}),document.getElementById("comments"))})}},{key:"new",value:function(){(new de).render(new je)}},{key:"edit",value:function(){var e=new de;je.find(this.params.id).then(function(t){return e.render(t)}),e.renderComments(this.params.id)}}])&&Ce(n.prototype,r),i&&Ce(n,i),t}();function ze(e){return(ze="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function De(e,t){return!t||"object"!==ze(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Ne(e){return(Ne=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Me(e,t){return(Me=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function Le(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var qe=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),De(this,Ne(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Me(e,t)}(t,o["Models"].Base),t}();Le(qe,"identity","Room.Member"),Le(qe,"resources",{url:"/user/rooms/:roomId/members",paginate:{per:100}}),Le(qe,"attributes",{}),Le(qe,"validate",[]);var Ve=qe;n(135),n(96);function Ue(e){return(Ue="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Fe(e,t){return!t||"object"!==Ue(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function We(e){return(We=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function He(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Je(e,t,n){return t&&He(e.prototype,t),n&&He(e,n),e}function Ye(e,t){return(Ye=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function $e(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var Ge=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Fe(this,We(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Ye(e,t)}(t,o["Models"].Base),Je(t,null,[{key:"receivedSignal",value:function(e){console.log("User.receivedSignal: ".concat(e))}}]),Je(t,[{key:"receivedSignal",value:function(e){console.log("User#receivedSignal: ".concat(e))}}]),t}();$e(Ge,"identity","User"),$e(Ge,"resources",{url:"/users",admin:{url:"/admin/users"}}),$e(Ge,"paginate",{per:10}),$e(Ge,"attributes",{email:{validations:{presence:!0,format:{with:/^([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})$/i}}},username:{validations:{presence:!0,format:{with:/^[a-z][a-z0-9_-]*$/i}}},password:{validations:{presence:{on:"create"},confirmation:!0}},passwordConfirmation:{remoteName:"password_confirmation"},confirmed:{},createdAt:{type:"Date",remoteName:"created_at"},updatedAt:{type:"Date",remoteName:"updated_at"}});var Ke=Ge;function Qe(e){return(Qe="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Xe(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Ze(e,t){return!t||"object"!==Qe(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function et(e){return(et=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function tt(e,t){return(tt=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var nt=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Ze(this,et(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&tt(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){this.connectWith(Ke)}},{key:"receivedSignal",value:function(e,t){switch(e){case"Room member_joined":this._memberJoined(t.room_id);break;case"Room member_left":this._memberLeft(t.room_id);break;case"Room created":var n=this._renderRoom(t.room);document.getElementById("rooms_list").insertAdjacentHTML("beforeend",n);break;case"Room destroyed":var r=document.getElementById("room_".concat(t.room_id));r.parentNode.removeChild(r)}}},{key:"_memberJoined",value:function(e){var t=this._membersNode(e);t.textContent=parseInt(t.text())+1}},{key:"_memberLeft",value:function(e){var t=this._membersNode(e);t.textContent=parseInt(t.text())-1}},{key:"_membersNode",value:function(e){document.querySelector("#room_".concat(e," td.members"))}},{key:"_renderRoom",value:function(e){"\n    <tr id='room_".concat(e.id,"'>\n      <td>").concat(e.name,"</td>\n      <td class='members'>0</td>\n      <td>\n        <a rel='nofollow' data-method='patch' href='/user/rooms/").concat(e.id,"/join'>Join</a> |\n        <a data-confirm='R U sure?' rel='nofollow' data-method='delete'\n          href='/user/rooms/").concat(e.id,"'>Destroy</a>\n      </td>\n    </tr>\n    ")}}])&&Xe(n.prototype,r),i&&Xe(n,i),t}();function rt(e){return(rt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function ot(e,t){return!t||"object"!==rt(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function it(e){return(it=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function at(e,t){return(at=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function ct(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var ut=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),ot(this,it(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&at(e,t)}(t,o["Models"].Base),t}();ct(ut,"identity","Room"),ct(ut,"resources",{}),ct(ut,"attributes",{}),ct(ut,"validate",[]);var lt=ut;function ft(e){return(ft="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function st(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function pt(e,t){return!t||"object"!==ft(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function yt(e){return(yt=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function mt(e,t){return(mt=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var dt=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=pt(this,yt(t).call(this,n))).roomId=n.id,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&mt(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){this.connectWith(lt),this._handleSendingMessage()}},{key:"renderMembers",value:function(e){var t=!0,n=!1,r=void 0;try{for(var o,i=e[Symbol.iterator]();!(t=(o=i.next()).done);t=!0){var a=o.value;this._memberJoined(a)}}catch(e){n=!0,r=e}finally{try{t||null==i.return||i.return()}finally{if(n)throw r}}}},{key:"receivedMessage",value:function(e,t){var n="<p><b>".concat(t,"</b>: ").concat(e,"</p>");document.getElementById("messages").insertAdjacentHTML("beforeend",n)}},{key:"receivedSignal",value:function(e,t){switch(e){case"Room member_joined":if(t.room_id!==this.roomId)return;this._memberJoined(t.member);break;case"Room member_left":if(t.room_id!==this.roomId)return;this._memberLeft(t.member)}}},{key:"_handleSendingMessage",value:function(){var e=this;document.querySelector("[data-behavior~=room-speaker]").addEventListener("keypress",function(t){13===t.keyCode&&(t.preventDefault(),o.Env.loco.emit({signal:"message",txt:t.target.value,room_id:e.roomId}),t.target.value="")})}},{key:"_memberJoined",value:function(e){var t="<li id='user_".concat(e.id,"'>").concat(e.username,"</li>");document.getElementById("members").insertAdjacentHTML("beforeend",t)}},{key:"_memberLeft",value:function(e){var t=document.querySelector("#members li#user_".concat(e.id));t.parentNode.removeChild(t)}}])&&st(n.prototype,r),i&&st(n,i),t}();function bt(e){return(bt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function ht(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function vt(e,t){return!t||"object"!==bt(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function wt(e){return(wt=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function gt(e,t){return(gt=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Ot=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),vt(this,wt(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&gt(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){}},{key:"index",value:function(){(new nt).render()}},{key:"show",value:function(){var e=new dt({id:this.params.id});this.setView("show",e),e.render(),Ve.all({roomId:this.params.id}).then(function(t){return e.renderMembers(t.resources)})}}])&&ht(n.prototype,r),i&&ht(n,i),t}();function _t(e){return(_t="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function jt(e,t){return!t||"object"!==_t(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function St(e){return(St=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Et(e,t){return(Et=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}Object.assign(o.Controllers.Base.prototype,p);var Pt=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),jt(this,St(t).apply(this,arguments))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Et(e,t)}(t,o["Controllers"].Base),t}();Object.assign(Pt,{Articles:Be,Rooms:Ot});var kt=Pt;function At(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}var It=function(){function e(){!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e)}var t,n,r;return t=e,(n=[{key:"receivedSignal",value:function(e){switch(e.signal){case"ping":this._pingSignal();break;case"message":this._messageSignal(e)}}},{key:"_pingSignal",value:function(){o.Env.namespaceController.constructor===kt&&alert("Ping!")}},{key:"_messageSignal",value:function(e){var t=this._getRoomView();t&&t.receivedMessage(e.message,e.author)}},{key:"_getRoomView",value:function(){return o.Env.namespaceController.constructor===kt&&(o.Env.controller.constructor===Ot&&("show"===o.Env.action&&o.Env.controller.getView("show")))}}])&&At(t.prototype,n),r&&At(t,r),e}(),Ct="PREPEND_ARTICLES",Tt="PREPEND_USERS";function xt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function Rt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}var Bt=Object(P.a)({articles:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case Ct:return[].concat(xt(t.payload.articles),xt(e));case h:return xt(t.payload.articles);case v:return[].concat(xt(e.slice(0,t.payload.index)),[t.payload.article],xt(e.slice(t.payload.index+1)));default:return e}},users:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case"SET_USERS":return Rt(t.payload.users);case Tt:return[].concat(Rt(t.payload.users),Rt(e));default:return e}}}),zt=Object(P.b)(Bt,{users:[],articles:[]});function Dt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function Nt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function Mt(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){Lt(e,t,n[t])})}return e}function Lt(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var qt=Object(P.a)({articles:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case b:return[].concat(Dt(e),Dt(t.payload.articles));case h:return Dt(t.payload.articles);case v:return[].concat(Dt(e.slice(0,t.payload.index)),[t.payload.article],Dt(e.slice(t.payload.index+1)));default:return e}},comments:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=arguments.length>1?arguments[1]:void 0;switch(t.type){case w:return Mt({},e,Lt({},t.payload.articleId,[].concat(Nt(e[t.payload.articleId]),Nt(t.payload.comments))));case g:return null==e[t.payload.articleId]?e:Mt({},e,Lt({},t.payload.articleId,e[t.payload.articleId].filter(function(e){return e.id!==t.payload.id})));case O:return Lt({},t.payload.articleId,t.payload.comments);case"UPDATE_COMMENT":var n=t.payload.articleId;return Mt({},e,Lt({},n,[].concat(Nt(e[n].slice(0,t.payload.index)),[t.payload.comment],Nt(e[n].slice(t.payload.index+1)))));default:return e}}}),Vt=Object(P.b)(qt,{articles:[],comments:{}}),Ut=function(e){return function(t,n){var r=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},o=t[e];r.parentId&&(o=o[r.parentId]);var i=o.find(function(e){return e.id===n});return i?[i,o.indexOf(i)]:[null,null]}},Ft=Ut("articles"),Wt=Ut("comments");function Ht(e){var t=e.comment,n=new o.Services.Date(t.createdAt).strftime("%d %b %y");return m.a.createElement("p",{id:"comment_".concat(t.id)},m.a.createElement("b",null,t.author)," on ",m.a.createElement("i",null,n)," (",m.a.createElement("a",{href:"/admin/articles/".concat(t.articleId,"/comments/").concat(t.id,"/edit")},"edit"),")",m.a.createElement("br",null),t.text)}Ht.propTypes={comment:Z.a.instanceOf(Q).isRequired};var Jt=Ht;function Yt(e){var t=e.comments;return 0===t.length?m.a.createElement("p",null,"No comments"):m.a.createElement(m.a.Fragment,null,t.map(function(e){return m.a.createElement(Jt,{comment:e,key:e.id})}))}Yt.propTypes={comments:Z.a.arrayOf(Z.a.instanceOf(Q)).isRequired};var $t=Yt;function Gt(e){return(Gt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Kt(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Qt(e,t){return!t||"object"!==Gt(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Xt(e){return(Xt=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Zt(e,t){return(Zt=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var en=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=Qt(this,Xt(t).call(this,n))).article=null,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Zt(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"receivedSignal",value:function(e){var t=this;switch(e){case"updated":this.article.reload().then(function(){t.article.applyChanges(),t._renderArticle()})}}},{key:"render",value:function(e){this.article=e,this.connectWith(e),this._renderArticle()}},{key:"renderComments",value:function(e){Object(d.render)(m.a.createElement($t,{comments:e}),document.getElementById("comments"))}},{key:"_renderArticle",value:function(){document.getElementById("article_author").textContent=this.article.author,document.getElementById("article_title").textContent=this.article.title,document.getElementById("article_text").textContent=this.article.content}}])&&Kt(n.prototype,r),i&&Kt(n,i),t}();function tn(e){return(tn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function nn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function rn(e,t){return!t||"object"!==tn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function on(e){return(on=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function an(e,t){return(an=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var cn=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),rn(this,on(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&an(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(e){e.setDefaultValuesForAdminReview(),new o.UI.Form({id:"edit_article_form",for:e}).render()}}])&&nn(n.prototype,r),i&&nn(n,i),t}();function un(e){var t=e.article,n=new o.Services.Date(t.publishedAt).toString("short");return m.a.createElement("article",{id:"article_".concat(t.id)},m.a.createElement("h2",null,t.title),m.a.createElement("p",null,m.a.createElement("i",null,t.author," wrote this on ",n," /"," ",m.a.createElement("span",{className:"comments_quantity"},t.commentsCount," comment",1===t.commentsCount?"":"s"))),m.a.createElement("p",null,t.content),m.a.createElement("p",null,m.a.createElement("a",{href:"/admin/articles/".concat(t.id,"/edit")},"Review")))}un.propTypes={article:Z.a.instanceOf(je).isRequired};var ln=un;function fn(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function sn(e){var t=fn(Object(y.useState)(e.articles),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=zt.subscribe(function(){return r(zt.getState().articles)});return function(){e()}});var o=n.map(function(e){return m.a.createElement(ln,{key:e.id,article:e})});return m.a.createElement(m.a.Fragment,null,o)}sn.propTypes={articles:Z.a.arrayOf(Z.a.instanceOf(je)).isRequired};var pn=sn;function yn(e){return(yn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function mn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function dn(e,t){return!t||"object"!==yn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function bn(e){return(bn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function hn(e,t){return(hn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var vn=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),dn(this,bn(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&hn(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"published",value:function(){je.get("published").then(function(e){zt.dispatch(j(e.resources)),Object(d.render)(m.a.createElement(pn,{articles:e.resources}),document.getElementById("articles"))})}},{key:"edit",value:function(){var e=new en;je.find(this.params.id).then(function(t){e.render(t),(new cn).render(t)}),Q.all({articleId:this.params.id}).then(function(t){e.renderComments(t.resources)})}}])&&mn(n.prototype,r),i&&mn(n,i),t}();function wn(e){return(wn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function gn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function On(e,t){return!t||"object"!==wn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function _n(e){return(_n=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function jn(e,t){return(jn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Sn=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),On(this,_n(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&jn(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=new o.UI.Form({for:new Q({id:e.commentId,resource:"admin"}),id:"edit_comment_".concat(e.commentId),initObj:!0});t.render(),o.Env.test={commentFormObj:t.getObj()}}}])&&gn(n.prototype,r),i&&gn(n,i),t}();function En(e){return(En="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Pn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function kn(e,t){return!t||"object"!==En(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function An(e){return(An=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function In(e,t){return(In=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Cn=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),kn(this,An(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&In(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"edit",value:function(){(new Sn).render({commentId:this.params.id})}}])&&Pn(n.prototype,r),i&&Pn(n,i),t}();function Tn(e){return(Tn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function xn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Rn(e,t){return!t||"object"!==Tn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Bn(e){return(Bn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function zn(e,t){return(zn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Dn=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Rn(this,Bn(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&zn(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({id:"sign_in_admin",delegator:this,callbackSuccess:"_signedIn"}).render()}},{key:"_signedIn",value:function(){window.location.href="/admin"}}])&&xn(n.prototype,r),i&&xn(n,i),t}();function Nn(e){return(Nn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Mn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Ln(e,t){return!t||"object"!==Nn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function qn(e){return(qn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Vn(e,t){return(Vn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Un=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Ln(this,qn(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Vn(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"new",value:function(){(new Dn).render()}}])&&Mn(n.prototype,r),i&&Mn(n,i),t}(),Fn=function(e){var t=e.user;return m.a.createElement("tr",{id:"user_".concat(t.id)},m.a.createElement("td",null,t.email),m.a.createElement("td",null,t.username),m.a.createElement("td",{className:"confirmed"},t.confirmed?"Yes":"No"),m.a.createElement("td",null,m.a.createElement("a",{href:"/admin/users/".concat(t.id)},"Show")," |"," ",m.a.createElement("a",{href:"/admin/users/".concat(t.id,"/edit")},"Edit")," |"," ",m.a.createElement("a",{href:"/admin/users/".concat(t.id),"data-method":"delete","data-confirm":"Are you sure?"},"Delete")," ","|"," ",m.a.createElement("a",{href:"#",onClick:function(e){return function(e,t){e.preventDefault(),o.Env.loco.emit({signal:"ping",user_id:t})}(e,t.id)}},"Ping")))};Fn.propTypes={user:Z.a.instanceOf(Ke).isRequired};var Wn=Fn;function Hn(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Jn(e){var t=Hn(Object(y.useState)(e.users),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=zt.subscribe(function(){return r(zt.getState().users)});return function(){e()}});var o=n.map(function(e){return m.a.createElement(Wn,{key:"user_".concat(e.id),user:e})});return m.a.createElement(m.a.Fragment,null,o)}Jn.propTypes={users:Z.a.arrayOf(Z.a.instanceOf(Ke)).isRequired};var Yn=Jn;function $n(e){return($n="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Gn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Kn(e,t){return!t||"object"!==$n(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Qn(e){return(Qn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Xn(e,t){return(Xn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Zn=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=Kn(this,Qn(t).call(this,n))).user=n.user,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Xn(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){document.getElementById("user_email").textContent=this.user.email,document.getElementById("user_username").textContent=this.user.username,document.getElementById("user_confirmed").textContent=this.user.confirmed?"Yes":"No",this._updateEditLink()}},{key:"_updateEditLink",value:function(){var e=document.getElementById("edit_link"),t=e.getAttribute("href");e.setAttribute("href",t.replace("/0/","/#{@user.id}/"))}}])&&Gn(n.prototype,r),i&&Gn(n,i),t}();function er(e){return(er="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function tr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function nr(e,t){return!t||"object"!==er(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function rr(e){return(rr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function or(e,t){return(or=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var ir=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=nr(this,rr(t).call(this,n))).user=n.user,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&or(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({for:this.user,initObj:!0,id:"admin_user_form"}).render()}}])&&tr(n.prototype,r),i&&tr(n,i),t}();function ar(e){return(ar="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function cr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function ur(e,t){return!t||"object"!==ar(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function lr(e){return(lr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function fr(e,t){return(fr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var sr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),ur(this,lr(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&fr(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"index",value:function(){Ke.get("all").then(function(e){var t;zt.dispatch((t=e.resources,{type:Tt,payload:{users:t}})),Object(d.render)(m.a.createElement(Yn,{users:e.resources}),document.querySelector("table tbody"))})}},{key:"show",value:function(){Ke.find(this.params.id).then(function(e){return new Zn({user:e}).render()})}},{key:"edit",value:function(){new ir({user:new Ke({id:this.params.id})}).render()}}])&&cr(n.prototype,r),i&&cr(n,i),t}();function pr(e){return(pr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function yr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function mr(e,t){return!t||"object"!==pr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function dr(e){return(dr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function br(e,t){return(br=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}Object.assign(o.Controllers.Base.prototype,p);var hr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),mr(this,dr(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&br(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){this.setScope("admin")}}])&&yr(n.prototype,r),i&&yr(n,i),t}();Object.assign(hr,{Articles:vn,Comments:Cn,Sessions:Un,Users:sr});var vr=hr;function wr(e){var t=e.comment,n=new o.Services.Date(t.createdAt).strftime("%d %b %y");return m.a.createElement("p",{id:"comment_".concat(t.id)},m.a.createElement("b",null,t.author)," on ",m.a.createElement("i",null,n)," wrote:",m.a.createElement("br",null),t.text)}wr.propTypes={comment:Z.a.instanceOf(Q).isRequired};var gr=wr;function Or(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function _r(e){var t=e.articleId,n=Or(Object(y.useState)(e.comments),2),r=n[0],o=n[1];Object(y.useEffect)(function(){var e=Vt.subscribe(function(){o(ee(Vt.getState(),t))});return function(){e()}});var i=r.map(function(e){return m.a.createElement(gr,{key:e.id,comment:e})});return 0===r.length?m.a.createElement("p",{id:"no_comments"},"No comments."):m.a.createElement(m.a.Fragment,null,i)}_r.propTypes={articleId:Z.a.number.isRequired,comments:Z.a.arrayOf(Z.a.instanceOf(Q)).isRequired};var jr=_r;function Sr(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Er(e){var t=Sr(Object(y.useState)(e.comments),2),n=t[0],r=t[1];return Object(y.useEffect)(function(){var t=Vt.subscribe(function(){r(ee(Vt.getState(),e.articleId))});return function(){t()}}),m.a.createElement(m.a.Fragment,null,"".concat(n.length," comment").concat(1===n.length?"":"s"))}Er.propTypes={articleId:Z.a.number.isRequired,comments:Z.a.arrayOf(Z.a.instanceOf(Q)).isRequired};var Pr=Er;function kr(e){return(kr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Ar(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Ir(e,t){return!t||"object"!==kr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Cr(e){return(Cr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Tr(e,t){return(Tr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var xr=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=Ir(this,Cr(t).call(this,n))).article=null,e.newComment=n.comment,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Tr(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({for:this.newComment,initObj:!0,id:"new_comment"}).render()}},{key:"renderArticle",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null;this.article||(this.article=e,this.connectWith(this.article,{receiver:"_articleReceivedSignal"})),document.getElementById("title").textContent=this.article.title,document.getElementById("author").textContent=this.article.author;var t=new o.Services.Date(this.article.publishedAt);document.getElementById("pub_date").textContent=t.toString("short");var n=document.getElementById("text");n.innerHTML="";var r=(new o.Helpers.Text).simpleFormat(this.article.content);n.insertAdjacentHTML("beforeend",r)}},{key:"_articleReceivedSignal",value:function(e){var t=this;switch(e){case"updating":new s({warning:"Author is currently editing article. Be aware of possible changes."}).render();break;case"updated":this.article.reload().then(function(){t.article.applyChanges(),t.renderArticle()})}}}])&&Ar(n.prototype,r),i&&Ar(n,i),t}();function Rr(e){return(Rr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Br(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function zr(e,t){return!t||"object"!==Rr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Dr(e){return(Dr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Nr(e,t){return(Nr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Mr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),zr(this,Dr(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Nr(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"show",value:function(){var e=this,t=new Q({articleId:this.params.id}),n=new xr({comment:t});n.render(),je.find(this.params.id).then(function(e){Vt.dispatch(j([e])),n.renderArticle(e)}),Q.get("count",{articleId:this.params.id}).then(function(t){Q.all({articleId:e.params.id,total:t.total}).then(function(t){Vt.dispatch(E(t,e.params.id)),Object(d.render)(m.a.createElement(jr,{articleId:e.params.id,comments:t}),document.getElementById("comments")),Object(d.render)(m.a.createElement(Pr,{articleId:e.params.id,comments:t}),document.getElementById("comments_count"))})})}}])&&Br(n.prototype,r),i&&Br(n,i),t}();function Lr(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function qr(){var e=Lr(Object(y.useState)(1),2),t=e[0],n=e[1],r=Lr(Object(y.useState)(!1),2),o=r[0],i=r[1];function a(e){e.preventDefault();var r=t+1;n(r),je.get("all",{page:r}).then(function(e){e.resources.length>0?Vt.dispatch(_(e.resources)):i(!0)}).catch(function(e){return alert("Invalid URL: ".concat(e))})}return o?m.a.createElement("span",null,"No more posts."):m.a.createElement("a",{href:"#",id:"load_more",onClick:a},"Load more…")}function Vr(e){var t=e.article,n=new o.Services.Date(t.publishedAt).toString("short");return m.a.createElement("article",{id:"article_".concat(t.id)},m.a.createElement("h2",null,t.title),m.a.createElement("p",null,m.a.createElement("i",null,t.author," wrote this on ",n," /"," ",m.a.createElement("a",{href:"/articles/".concat(t.id,"#comments"),className:"comments_quantity"},t.commentsCount," comment",1===t.commentsCount?"":"s"))),m.a.createElement("p",null,t.content),m.a.createElement("p",null,m.a.createElement("a",{href:"/articles/".concat(t.id)},"Continued…")))}Vr.propTypes={article:Z.a.instanceOf(je).isRequired};var Ur=Vr;function Fr(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Wr(e){var t=Fr(Object(y.useState)(e.articles),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=Vt.subscribe(function(){return r(Vt.getState().articles)});return function(){e()}});var o=n.map(function(e){return m.a.createElement(Ur,{key:e.id,article:e})});return m.a.createElement(m.a.Fragment,null,o)}Wr.propTypes={articles:Z.a.arrayOf(Z.a.instanceOf(je)).isRequired};var Hr=Wr;function Jr(e){return(Jr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Yr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function $r(e,t){return!t||"object"!==Jr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Gr(e){return(Gr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Kr(e,t){return(Kr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Qr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),$r(this,Gr(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Kr(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"index",value:function(){Object(d.render)(m.a.createElement(qr,null),document.getElementById("load_more_wrapper")),je.get("all",{page:1}).then(function(e){Vt.dispatch(j(e.resources)),Object(d.render)(m.a.createElement(Hr,{articles:e.resources}),document.getElementById("articles"))})}}])&&Yr(n.prototype,r),i&&Yr(n,i),t}();function Xr(e){return(Xr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Zr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function eo(e,t){return!t||"object"!==Xr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function to(e){return(to=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function no(e,t){return(no=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var ro=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),eo(this,to(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&no(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({for:new Ke,delegator:this,callbackSuccess:"_created"}).render()}},{key:"receivedSignal",value:function(e){switch(e){case"confirming":this._confirming();break;case"confirmed":this._confirmed()}}},{key:"_created",value:function(e){this.connectWith(new Ke({id:e.id})),document.querySelector("form").style.display="none",document.getElementById("sign_in_paragraph").classList.remove("none"),document.getElementById("verification_info").classList.remove("none"),new s({notice:e.notice}).render()}},{key:"_confirming",value:function(){document.getElementById("verification_info").textContent=document.getElementById("verification_progress").textContent}},{key:"_confirmed",value:function(){window.location.href="/user/sessions/new?event=confirmed"}}])&&Zr(n.prototype,r),i&&Zr(n,i),t}();function oo(e){return(oo="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function io(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function ao(e,t){return!t||"object"!==oo(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function co(e){return(co=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function uo(e,t){return(uo=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var lo=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),ao(this,co(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&uo(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"new",value:function(){(new ro).render()}}])&&io(n.prototype,r),i&&io(n,i),t}();function fo(e){return(fo="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function so(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function po(e,t){return!t||"object"!==fo(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function yo(e){return(yo=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function mo(e,t){return(mo=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}Object.assign(o.Controllers.Base.prototype,p);var bo=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),po(this,yo(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&mo(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){this.setScope("main")}}])&&so(n.prototype,r),i&&so(n,i),t}();Object.assign(bo,{Articles:Mr,Pages:Qr,Users:lo});var ho=bo;function vo(e){return(vo="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function wo(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function go(e,t){return!t||"object"!==vo(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Oo(e){return(Oo=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function _o(e,t){return(_o=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function jo(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function So(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}var Eo=function(e){var t=e.id;o.Env.namespaceController.constructor===vr?je.find({id:t,abbr:!0,resource:"admin"}).then(function(e){zt.dispatch({type:Ct,payload:{articles:[e]}})}):je.find({id:t,abbr:!0}).then(function(e){return Vt.dispatch(_([e]))})},Po=function(e,t){var n=e.article_id,r=Vt;o.Env.namespaceController.constructor===vr?r=zt:o.Env.namespaceController.constructor===kt&&(r=R);var i=So(Ft(r.getState(),n),2),a=i[0],c=i[1];if(a){var u=new je(function(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){jo(e,t,n[t])})}return e}({},a,{commentsCount:a.commentsCount+t}));r.dispatch(S(u,c))}},ko=function(e){var t=e.article_id,n=e.id,r=R,i={articleId:t,id:n};o.Env.namespaceController.constructor===ho&&(r=Vt,i.resource="main"),So(Ft(r.getState(),t),1)[0]&&Q.find(i).then(function(e){return r.dispatch(function(e,t){return{type:w,payload:{articleId:t,comments:e}}}([e],t))})},Ao=function(e){var t=e.article_id,n=e.id,r=Vt;o.Env.namespaceController.constructor===kt&&(r=R),r.dispatch(function(e,t){return{type:g,payload:{id:e,articleId:t}}}(n,t))},Io=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),go(this,Oo(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&_o(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"receivedSignal",value:function(e,t){switch(e){case"Article created":n=t.id,o.Env.namespaceController.constructor===kt&&je.find({id:n,abbr:!0}).then(function(e){return R.dispatch(_([e]))});break;case"Article published":Eo(t);break;case"Article updated":!function(e){var t=e.id,n={id:t,abbr:!0},r=Vt;o.Env.namespaceController.constructor===vr?(n.resource="admin",r=zt):o.Env.namespaceController.constructor===kt&&(r=R);var i=So(Ft(r.getState(),t),2),a=i[0],c=i[1];a&&je.find(n).then(function(e){return r.dispatch(S(e,c))})}(t);break;case"Article.Comment created":Po(t,1),ko(t);break;case"Article.Comment destroyed":Po(t,-1),Ao(t);break;case"Article.Comment updated":!function(e){var t=e.article_id,n=e.id,r=Vt;o.Env.namespaceController.constructor===kt&&(r=R);var i=So(Wt(r.getState(),n,{parentId:t}),2),a=i[0],c=i[1];a&&a.reload().then(function(){a.applyChanges(),r.dispatch({type:"UPDATE_COMMENT",payload:{comment:a,index:c,articleId:t}})})}(t);break;case"User created":Ke.find(t.id).then(function(e){return zt.dispatch({type:Tt,payload:{users:[e]}})})}var n}},{key:"call",value:function(){this.connectWith([je,Q,Ke])}}])&&wo(n.prototype,r),i&&wo(n,i),t}();o.Deps.cable=Object(i.createConsumer)(),o.Deps.NotificationCenter=It;var Co=new Io;new o.Loco({notifications:{enable:!0,log:!0,size:10},postInit:function(){Co.call(),"test"===document.querySelector("body").getAttribute("data-rails-env")&&o.Env.loco.getWire().setPollingTime(1e3)}}).init();n(138);function To(e){return(To="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function xo(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Ro(e,t){return!t||"object"!==To(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Bo(e){return(Bo=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function zo(e,t){return(zo=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Do,No,Mo,Lo=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Ro(this,Bo(t).call(this))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&zo(e,t)}(t,o["Validators"].Base),n=t,(r=[{key:"validate",value:function(){if(null!=this.val)switch(To(this.val)){case"string":var e=this._getVulgarWord();if(!new RegExp("".concat(e),["i"]).exec(this.val))return;this._addErrorMessage();break;default:throw new TypeError("Vulgarity validator is applicable only for strings and ".concat(this.attr," isn't."))}}},{key:"_addErrorMessage",value:function(){var e=o.I18n[o.Env.loco.getLocale()].errors.messages.vulgarity;this.obj.addErrorMessage(e,{for:this.attr})}},{key:"_getVulgarWord",value:function(){switch(o.Env.loco.getLocale()){case"pl":return"kurwa";case"en":return"fuck"}}}])&&xo(n.prototype,r),i&&xo(n,i),t}();Mo="Vulgarity",(No="identity")in(Do=Lo)?Object.defineProperty(Do,No,{value:Mo,enumerable:!0,configurable:!0,writable:!0}):Do[No]=Mo;var qo=Lo;Object.assign(o.Validators,{Vulgarity:qo}),o.I18n.pl={variants:{few:function(e){var t=parseInt(o.Utils.String.last(String(e)));-1!==[2,3,4].indexOf(t)&&(2!==String(e).length||String(e)[0])}},models:{},attributes:{},ui:{form:{sending:"Wysyłam...",success:"Sukces",errors:{connection:"Błąd z połączeniem",invalid_data:"Nieprawidłowe dane"}}},date:{formats:{default:"%d-%m-%Y",short:"%d %b",long:"%B %d, %Y"},day_names:["niedziela","poniedziałek","wtorek","środa","czwartek","piątek","sobota"],abbr_day_names:["nie","pon","wto","śro","czw","pią","sob"],month_names:["styczeń","luty","marzec","kwiecień","maj","czerwiec","lipiec","sierpień","wrzesień","październik","listopad","grudzień"],abbr_month_names:["sty","lut","mar","kwi","maj","cze","lip","sie","wrz","paź","lis","gru"]},errors:{messages:{accepted:"musi zostać zaakceptowane",blank:"nie może być puste",confirmation:"nie zgadza się z polem %{attribute}",empty:"nie może być puste",equal_to:"musi być równe %{count}",even:"musi być parzyste",exclusion:"jest zarezerwowane",greater_than:"musi być większe od %{count}",greater_than_or_equal_to:"musi być większe lub równe %{count}",inclusion:"nie znajduje się na liście dopuszczalnych wartości",invalid:"jest nieprawidłowe",less_than:"musi być mniejsze od %{count}",less_than_or_equal_to:"musi być mniejsze lub równe %{count}",not_a_number:"nie jest liczbą",not_an_integer:"musi być liczbą całkowitą",odd:"musi być nieparzyste",present:"musi być puste",too_long:{few:"jest za długie (maksymalnie %{count} znaki)",many:"jest za długie (maksymalnie %{count} znaków)",one:"jest za długie (maksymalnie jeden znak)",other:"jest za długie (maksymalnie %{count} znaków)"},too_short:{few:"jest za krótkie (przynajmniej %{count} znaki)",many:"jest za krótkie (przynajmniej %{count} znaków)",one:"jest za krótkie (przynajmniej jeden znak)",other:"jest za krótkie (przynajmniej %{count} znaków)"},wrong_length:{few:"ma nieprawidłową długość (powinna wynosić %{count} znaki)",many:"ma nieprawidłową długość (powinna wynosić %{count} znaków)",one:"ma nieprawidłową długość (powinna wynosić jeden znak)",other:"ma nieprawidłową długość (powinna wynosić %{count} znaków)"},other_than:"musi być inna niż %{count}"}}},o.I18n.pl.models.User="Użytkownik",o.I18n.pl.attributes.User={password:"Hasło",passwordConfirmation:"Powtórz hasło"},o.I18n.en.errors.messages.vulgarity="contains strong language.",o.I18n.pl.errors.messages.vulgarity="zawiera mocny język.",Object.assign(o.Controllers,{Admin:vr,Main:ho,User:kt}),Object.assign(je,{Comment:Q}),Object.assign(lt,{Member:Ve}),Object.assign(o.Models,{Article:je,Room:lt,User:Ke}),window.test={Channels:o.Channels,Env:o.Env}}});
+!function(e){function t(t){for(var r,a,c=t[0],u=t[1],l=t[2],s=0,p=[];s<c.length;s++)a=c[s],o[a]&&p.push(o[a][0]),o[a]=0;for(r in u)Object.prototype.hasOwnProperty.call(u,r)&&(e[r]=u[r]);for(f&&f(t);p.length;)p.shift()();return i.push.apply(i,l||[]),n()}function n(){for(var e,t=0;t<i.length;t++){for(var n=i[t],r=!0,c=1;c<n.length;c++){var u=n[c];0!==o[u]&&(r=!1)}r&&(i.splice(t--,1),e=a(a.s=n[0]))}return e}var r={},o={0:0},i=[];function a(t){if(r[t])return r[t].exports;var n=r[t]={i:t,l:!1,exports:{}};return e[t].call(n.exports,n,n.exports,a),n.l=!0,n.exports}a.m=e,a.c=r,a.d=function(e,t,n){a.o(e,t)||Object.defineProperty(e,t,{enumerable:!0,get:n})},a.r=function(e){"undefined"!=typeof Symbol&&Symbol.toStringTag&&Object.defineProperty(e,Symbol.toStringTag,{value:"Module"}),Object.defineProperty(e,"__esModule",{value:!0})},a.t=function(e,t){if(1&t&&(e=a(e)),8&t)return e;if(4&t&&"object"==typeof e&&e&&e.__esModule)return e;var n=Object.create(null);if(a.r(n),Object.defineProperty(n,"default",{enumerable:!0,value:e}),2&t&&"string"!=typeof e)for(var r in e)a.d(n,r,function(t){return e[t]}.bind(null,r));return n},a.n=function(e){var t=e&&e.__esModule?function(){return e.default}:function(){return e};return a.d(t,"a",t),t},a.o=function(e,t){return Object.prototype.hasOwnProperty.call(e,t)},a.p="";var c=window.webpackJsonp=window.webpackJsonp||[],u=c.push.bind(c);c.push=t,c=c.slice();for(var l=0;l<c.length;l++)t(c[l]);var f=u;i.push([143,1]),n()}({101:function(e,t,n){},102:function(e,t,n){},143:function(e,t,n){"use strict";n.r(t);n(101),n(102);var r=n(98);n.n(r).a.start();var o=n(1),i=n(99);n(4),n(5),n(6),n(3),n(34),n(10),n(11),n(2),n(7),n(8);function a(e){return(a="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function c(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function u(e,t){return!t||"object"!==a(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function l(e){return(l=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function f(e,t){return(f=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var s=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=u(this,l(t).call(this,n))).notice=n.notice,e.alert=n.alert,e.warning=n.warning,e.hide=n.hide,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&f(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"setNotice",value:function(e){this.notice=e}},{key:"setAlert",value:function(e){this.alert=e}},{key:"setWarning",value:function(e){this.warning=e}},{key:"render",value:function(){var e=document.querySelector(".flash");e.classList.remove("notice"),e.classList.remove("alert"),e.classList.remove("warning"),null!=this.notice?(e.classList.add("notice"),document.querySelector(".flash p").textContent=this.notice):null!=this.alert?(e.classList.add("alert"),document.querySelector(".flash p").textContent=this.alert):null!=this.warning&&(e.classList.add("warning"),document.querySelector(".flash p").textContent=this.warning),e.classList.remove("none"),this.hide&&this.hideAfterTime()}},{key:"hideAfterTime",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:4e3;setTimeout(function(){document.querySelector(".flash").classList.add("none")},e)}}])&&c(n.prototype,r),i&&c(n,i),t}(),p={disconnectedForTooLong:function(){new s({alert:"You have been disconnected from the server for too long. Reload page!",hide:!1}).render()}},y=(n(30),n(0)),m=n.n(y),d=n(16),b="ADD_ARTICLES",h="PREPEND_ARTICLES",v="REMOVE_ARTICLE",w="SET_ARTICLES",g="UPDATE_ARTICLE",O="ADD_COMMENTS",_="REMOVE_COMMENT",j="SET_COMMENTS",S="UPDATE_COMMENT",E="PREPEND_USERS";function P(e){return{type:b,payload:{articles:e}}}function k(e){return{type:w,payload:{articles:e}}}function A(e,t){return{type:g,payload:{article:e,index:t}}}function I(e,t){return{type:j,payload:{comments:e,articleId:t}}}function C(e,t,n){return{type:S,payload:{comment:e,articleId:t,index:n}}}var T=n(20);n(19),n(41),n(31),n(42),n(25);function x(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}n(70),n(55),n(56),n(57);function R(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function B(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){z(e,t,n[t])})}return e}function z(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var D=Object(T.a)({articles:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case b:return[].concat(x(e),x(t.payload.articles));case v:return e.filter(function(e){return e.id!==t.payload.id});case w:return x(t.payload.articles);case g:return[].concat(x(e.slice(0,t.payload.index)),[t.payload.article],x(e.slice(t.payload.index+1)));default:return e}},comments:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=arguments.length>1?arguments[1]:void 0;switch(t.type){case O:return B({},e,z({},t.payload.articleId,[].concat(R(e[t.payload.articleId]),R(t.payload.comments))));case _:return null==e[t.payload.articleId]?e:B({},e,z({},t.payload.articleId,e[t.payload.articleId].filter(function(e){return e.id!==t.payload.id})));case j:return z({},t.payload.articleId,t.payload.comments);case S:var n=t.payload.articleId,r=t.payload.index;if(!r){var o=e[n].find(function(e){return e.id===t.payload.comment.id});r=e[n].indexOf(o)}return B({},e,z({},n,[].concat(R(e[n].slice(0,r)),[t.payload.comment],R(e[n].slice(r+1)))));default:return e}}}),N=Object(T.b)(D,{articles:[],comments:{}});function L(e){return(L="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function q(e,t){return!t||"object"!==L(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function M(e){return(M=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function V(e,t){return(V=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var U=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),q(this,M(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&V(e,t)}(t,o["Views"].Base),t}();n(94);function F(e){return(F="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function W(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function H(e,t){return!t||"object"!==F(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function J(e){return(J=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Y(e,t){return(Y=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var $=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=H(this,J(t).call(this,n))).article=null,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Y(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"renderArticle",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null;null==this.article&&(this.connectWith(e,{receiver:"articleReceivedSignal"}),this.article=e),document.getElementById("article_title").textContent=this.article.title,document.getElementById("article_text").textContent=this.article.content;var t=document.getElementById("publish_article");t&&null!=this.article.publishedAt?t.style.display="none":t&&(t.style.display=""),this._handlePublishing(),this._updateEditLink()}},{key:"articleReceivedSignal",value:function(e){var t=this;switch(e){case"updated":this.article.reload().then(function(){t.article.applyChanges(),t.renderArticle()});break;case"destroyed":window.location.href="/user/articles?message=deleted"}}},{key:"_handlePublishing",value:function(){var e=this;document.getElementById("publish_article").addEventListener("click",function(t){t.preventDefault(),t.target.textContent="Publishing...",e.article.put("publish").then(function(){return document.getElementById("publish_article").outerHTML="<span>Published!</span>"}).catch(function(){t.target.textContent="Publish",new s({alert:"Connection error!"}).render()})})}},{key:"_updateEditLink",value:function(){var e=document.getElementById("edit_link").getAttribute("href");document.getElementById("edit_link").setAttribute("href",e.replace("/0/","/".concat(this.article.id,"/")))}}])&&W(n.prototype,r),i&&W(n,i),t}();n(129),n(131);function G(e){return(G="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function K(e,t){return!t||"object"!==G(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Q(e){return(Q=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function X(e,t){return(X=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function Z(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var ee=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),K(this,Q(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&X(e,t)}(t,o["Models"].Base),t}();Z(ee,"identity","Article.Comment"),Z(ee,"remoteName","Comment"),Z(ee,"resources",{url:"/user/articles/:articleId/comments",paginate:{per:10},main:{url:"/articles/:articleId/comments",paginate:{per:5,param:"page-num"}},admin:{url:"/admin/articles/:articleId/comments",paginate:{per:5}}}),Z(ee,"attributes",{author:{validations:{presence:!0}},text:{validations:{presence:!0,vulgarity:!0}},articleId:{type:"Int",validations:{presence:!0},remoteName:"article_id"},createdAt:{type:"Date",remoteName:"created_at"},updatedAt:{type:"Date",remoteName:"updated_at"},emotion:{type:"Int"},pinned:{type:"Boolean"},adminRate:{type:"Int",remoteName:"admin_rate"},approved:{type:"Boolean"}});var te=ee,ne=(n(32),n(9)),re=n.n(ne),oe=function(e,t){return e.comments[t]};function ie(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function ae(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function ce(e){var t,n=e.comment,r=e.isAdmin,i=void 0!==r&&r,a=ae(Object(y.useState)(!1),2),c=a[0],u=a[1],l=new o.Services.Date(n.createdAt).strftime("%d %b %y"),f=function(e){e.preventDefault(),u(!0),n.approved=!0,n.updateAttribute("approved").then(function(e){e.ok&&N.dispatch(C(new te(function(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){ie(e,t,n[t])})}return e}({},n,{approved:!0})),n.articleId))})};return m.a.createElement("p",{id:"comment_".concat(n.id)},m.a.createElement("b",null,n.author)," on ",m.a.createElement("i",null,l),i?(t=n.approved?m.a.createElement(m.a.Fragment,null,m.a.createElement("span",null,"approved")," |"," "):c?m.a.createElement(m.a.Fragment,null,m.a.createElement("span",null,"approving...")," |"," "):m.a.createElement(m.a.Fragment,null,m.a.createElement("a",{href:"/user/articles/".concat(n.articleId,"/comments/").concat(n.id,"/approve"),className:"approve",onClick:f},"approve")," ","|"," "),m.a.createElement(m.a.Fragment,null," ","(",t,m.a.createElement("a",{href:"/user/articles/".concat(n.articleId,"/comments/").concat(n.id,"/edit")},"edit")," ","|"," ",m.a.createElement("a",{href:"/user/articles/".concat(n.articleId,"/comments/").concat(n.id),"data-method":"delete","data-confirm":"Are you sure?"},"delete"),")")):"",m.a.createElement("br",null),n.text)}ce.propTypes={comment:re.a.instanceOf(te).isRequired,isAdmin:re.a.bool};var ue=ce;function le(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function fe(e){var t=e.articleId,n=e.isAdmin||!1,r=le(Object(y.useState)(e.comments),2),o=r[0],i=r[1];if(Object(y.useEffect)(function(){var e=N.subscribe(function(){return i(oe(N.getState(),t))});return function(){e()}}),0===o.length)return m.a.createElement("p",null,"No comments.");var a=o.map(function(e){return m.a.createElement(ue,{key:e.id,comment:e,isAdmin:n})});return m.a.createElement(m.a.Fragment,null,a)}fe.propTypes={articleId:re.a.number.isRequired,isAdmin:re.a.bool,comments:re.a.arrayOf(re.a.instanceOf(te)).isRequired};var se=fe;function pe(e){return(pe="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function ye(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function me(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function de(e,t){return!t||"object"!==pe(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function be(e){return(be=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function he(e,t){return(he=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var ve=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=de(this,be(t).call(this,n))).article=null,e.form=null,e.changes=null,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&he(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(e){N.dispatch(P([e])),this.article=e,this.connectWith(this.article),this._handleApplyingChanges(),this.form=new o.UI.Form({for:this.article}),this.form.render()}},{key:"renderComments",value:function(e){te.all({articleId:e}).then(function(t){N.dispatch(I(t.resources,e)),Object(d.render)(m.a.createElement(se,{articleId:e,comments:t.resources,isAdmin:!0}),document.getElementById("comments"))})}},{key:"receivedSignal",value:function(e,t){var n=this;switch(e){case"updating":if(document.querySelector("h1").getAttribute("data-mark")!==t.mark)new s({warning:"Uuups someone else started editing this article."}).render();break;case"updated":this.article.reload().then(function(){n.changes=n.article.changes(),n._displayChanges()});break;case"destroyed":window.location.href="/user/articles?message=deleted"}}},{key:"_displayChanges",value:function(){for(var e=0,t=Object.entries(this.changes);e<t.length;e++){var n=ye(t[e],1)[0],r=document.querySelector("a.apply_changes[data-for=".concat(this.article.getAttrRemoteName(n),"]"));r&&r.classList.remove("none")}}},{key:"_handleApplyingChanges",value:function(){for(var e=this,t=0,n=Array.from(document.querySelectorAll("a.apply_changes"));t<n.length;t++){n[t].addEventListener("click",function(t){t.preventDefault();var n=e.article.getAttrName(t.target.getAttribute("data-for"));e.article[n]=e.changes[n].is,e.form.fill(n),t.target.classList.add("none")})}}}])&&me(n.prototype,r),i&&me(n,i),t}();function we(e){return(we="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function ge(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Oe(e,t){return!t||"object"!==we(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function _e(e){return(_e=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function je(e,t){return(je=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function Se(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var Ee=function(e){function t(e){var n;return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(n=Oe(this,_e(t).call(this,e))).published=null!=n.publishedAt,n}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&je(e,t)}(t,o["Models"].Base),n=t,(r=[{key:"vulgarityLevel",value:function(){(null!=this.title&&/fuck/i.exec(this.title)||null!=this.content&&/fuck/i.exec(this.content))&&this.addErrorMessage("Article contains strong language.",{for:"base"})}},{key:"setDefaultValuesForAdminReview",value:function(){this.adminRate=null==this.adminRate?3:this.adminRate,this.categoryId=null==this.categoryId?6:this.categoryId,this.adminReviewStartedAt=Date.now()}}])&&ge(n.prototype,r),i&&ge(n,i),t}();Se(Ee,"identity","Article"),Se(Ee,"resources",{url:"/user/articles",paginate:{per:5},main:{url:"/articles",paginate:{per:3}},admin:{url:"/admin/articles",paginate:{per:4}}}),Se(Ee,"attributes",{title:{validations:{presence:!0,length:{within:[3,255]}}},content:{validations:{presence:!0,length:{minimum:100}},remoteName:"text"},createdAt:{type:"Date",remoteName:"created_at"},updatedAt:{type:"Date",remoteName:"updated_at"},commentsCount:{type:"Int",remoteName:"comments_count"},publishedAt:{type:"Date",remoteName:"published_at"},published:{},adminReview:{remoteName:"admin_review"},adminRate:{type:"Int",remoteName:"admin_rate"},categoryId:{type:"Int",remoteName:"category_id"},adminReviewStartedAt:{remoteName:"admin_review_started_at"}}),Se(Ee,"validate",["vulgarityLevel"]);var Pe=Ee;function ke(e){var t=e.article,n=e.onArticleDestroyed;return m.a.createElement("tr",{id:"article_".concat(t.id)},m.a.createElement("td",null,t.title),m.a.createElement("td",null,t.content),m.a.createElement("td",{className:"comments_quantity"},t.commentsCount),m.a.createElement("td",{className:"published"},t.publishedAt?"yes":"no"),m.a.createElement("td",null,m.a.createElement("a",{href:"/user/articles/".concat(t.id)},"Show")," |",m.a.createElement("a",{href:"/user/articles/".concat(t.id,"/edit")},"Edit")," |",m.a.createElement("a",{href:"/user/articles/".concat(t.id),className:"delete_article",onClick:function(e){e.preventDefault(),confirm("Are you sure?")&&t.delete(null).then(function(e){var t;e.success&&N.dispatch((t=e.id,{type:v,payload:{id:t}})),n(e)})}},"Delete")))}ke.propTypes={article:re.a.instanceOf(Pe).isRequired,onArticleDestroyed:re.a.func.isRequired};var Ae=ke;function Ie(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Ce(e){var t=Ie(Object(y.useState)(e.articles),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=N.subscribe(function(){return r(N.getState().articles)});return function(){e()}});var o=n.map(function(t){return m.a.createElement(Ae,{key:t.id,article:t,onArticleDestroyed:e.onArticleDestroyed})});return m.a.createElement(m.a.Fragment,null,o)}Ce.propTypes={articles:re.a.arrayOf(re.a.instanceOf(Pe)).isRequired,onArticleDestroyed:re.a.func.isRequired};var Te=Ce;function xe(e){return(xe="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Re(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Be(e,t){return!t||"object"!==xe(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function ze(e){return(ze=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function De(e,t){return(De=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Ne=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Be(this,ze(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&De(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){this.layout=new U}},{key:"onArticleDestroyed",value:function(e){var t=new s;e.success?t.setNotice(e.notice):t.setAlert(e.alert),t.render()}},{key:"index",value:function(){var e=this;"deleted"===this.params.message&&new s({alert:"Article has been deleted."}).render();Pe.get("all").then(function(t){N.dispatch(k(t.resources)),Object(d.render)(m.a.createElement(Te,{articles:t.resources,onArticleDestroyed:e.onArticleDestroyed}),document.getElementById("article_list"))})}},{key:"show",value:function(){var e=this;this.showView=new $,Pe.find(this.params.id).then(function(t){N.dispatch(k([t])),e.showView.renderArticle(t)}),te.all({articleId:this.params.id}).then(function(t){N.dispatch(I(t.resources,e.params.id)),Object(d.render)(m.a.createElement(se,{articleId:e.params.id,comments:t.resources}),document.getElementById("comments"))})}},{key:"new",value:function(){(new ve).render(new Pe)}},{key:"edit",value:function(){var e=new ve;Pe.find(this.params.id).then(function(t){return e.render(t)}),e.renderComments(this.params.id)}}])&&Re(n.prototype,r),i&&Re(n,i),t}();function Le(e){return(Le="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function qe(e,t){return!t||"object"!==Le(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Me(e){return(Me=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Ve(e,t){return(Ve=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function Ue(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var Fe=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),qe(this,Me(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Ve(e,t)}(t,o["Models"].Base),t}();Ue(Fe,"identity","Room.Member"),Ue(Fe,"resources",{url:"/user/rooms/:roomId/members",paginate:{per:100}}),Ue(Fe,"attributes",{}),Ue(Fe,"validate",[]);var We=Fe;n(135),n(96);function He(e){return(He="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Je(e,t){return!t||"object"!==He(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Ye(e){return(Ye=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function $e(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Ge(e,t,n){return t&&$e(e.prototype,t),n&&$e(e,n),e}function Ke(e,t){return(Ke=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function Qe(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var Xe=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Je(this,Ye(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Ke(e,t)}(t,o["Models"].Base),Ge(t,null,[{key:"receivedSignal",value:function(e){console.log("User.receivedSignal: ".concat(e))}}]),Ge(t,[{key:"receivedSignal",value:function(e){console.log("User#receivedSignal: ".concat(e))}}]),t}();Qe(Xe,"identity","User"),Qe(Xe,"resources",{url:"/users",admin:{url:"/admin/users"}}),Qe(Xe,"paginate",{per:10}),Qe(Xe,"attributes",{email:{validations:{presence:!0,format:{with:/^([^@\s]+)@((?:[-a-z0-9]+\.)+[a-z]{2,})$/i}}},username:{validations:{presence:!0,format:{with:/^[a-z][a-z0-9_-]*$/i}}},password:{validations:{presence:{on:"create"},confirmation:!0}},passwordConfirmation:{remoteName:"password_confirmation"},confirmed:{},createdAt:{type:"Date",remoteName:"created_at"},updatedAt:{type:"Date",remoteName:"updated_at"}});var Ze=Xe;function et(e){return(et="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function tt(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function nt(e,t){return!t||"object"!==et(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function rt(e){return(rt=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function ot(e,t){return(ot=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var it=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),nt(this,rt(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&ot(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){this.connectWith(Ze)}},{key:"receivedSignal",value:function(e,t){switch(e){case"Room member_joined":this._memberJoined(t.room_id);break;case"Room member_left":this._memberLeft(t.room_id);break;case"Room created":var n=this._renderRoom(t.room);document.getElementById("rooms_list").insertAdjacentHTML("beforeend",n);break;case"Room destroyed":var r=document.getElementById("room_".concat(t.room_id));r.parentNode.removeChild(r)}}},{key:"_memberJoined",value:function(e){var t=this._membersNode(e);t.textContent=parseInt(t.text())+1}},{key:"_memberLeft",value:function(e){var t=this._membersNode(e);t.textContent=parseInt(t.text())-1}},{key:"_membersNode",value:function(e){document.querySelector("#room_".concat(e," td.members"))}},{key:"_renderRoom",value:function(e){"\n    <tr id='room_".concat(e.id,"'>\n      <td>").concat(e.name,"</td>\n      <td class='members'>0</td>\n      <td>\n        <a rel='nofollow' data-method='patch' href='/user/rooms/").concat(e.id,"/join'>Join</a> |\n        <a data-confirm='R U sure?' rel='nofollow' data-method='delete'\n          href='/user/rooms/").concat(e.id,"'>Destroy</a>\n      </td>\n    </tr>\n    ")}}])&&tt(n.prototype,r),i&&tt(n,i),t}();function at(e){return(at="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function ct(e,t){return!t||"object"!==at(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function ut(e){return(ut=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function lt(e,t){return(lt=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function ft(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var st=function(e){function t(e){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),ct(this,ut(t).call(this,e))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&lt(e,t)}(t,o["Models"].Base),t}();ft(st,"identity","Room"),ft(st,"resources",{}),ft(st,"attributes",{}),ft(st,"validate",[]);var pt=st;function yt(e){return(yt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function mt(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function dt(e,t){return!t||"object"!==yt(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function bt(e){return(bt=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function ht(e,t){return(ht=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var vt=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=dt(this,bt(t).call(this,n))).roomId=n.id,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&ht(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){this.connectWith(pt),this._handleSendingMessage()}},{key:"renderMembers",value:function(e){var t=!0,n=!1,r=void 0;try{for(var o,i=e[Symbol.iterator]();!(t=(o=i.next()).done);t=!0){var a=o.value;this._memberJoined(a)}}catch(e){n=!0,r=e}finally{try{t||null==i.return||i.return()}finally{if(n)throw r}}}},{key:"receivedMessage",value:function(e,t){var n="<p><b>".concat(t,"</b>: ").concat(e,"</p>");document.getElementById("messages").insertAdjacentHTML("beforeend",n)}},{key:"receivedSignal",value:function(e,t){switch(e){case"Room member_joined":if(t.room_id!==this.roomId)return;this._memberJoined(t.member);break;case"Room member_left":if(t.room_id!==this.roomId)return;this._memberLeft(t.member)}}},{key:"_handleSendingMessage",value:function(){var e=this;document.querySelector("[data-behavior~=room-speaker]").addEventListener("keypress",function(t){13===t.keyCode&&(t.preventDefault(),o.Env.loco.emit({signal:"message",txt:t.target.value,room_id:e.roomId}),t.target.value="")})}},{key:"_memberJoined",value:function(e){var t="<li id='user_".concat(e.id,"'>").concat(e.username,"</li>");document.getElementById("members").insertAdjacentHTML("beforeend",t)}},{key:"_memberLeft",value:function(e){var t=document.querySelector("#members li#user_".concat(e.id));t.parentNode.removeChild(t)}}])&&mt(n.prototype,r),i&&mt(n,i),t}();function wt(e){return(wt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function gt(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Ot(e,t){return!t||"object"!==wt(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function _t(e){return(_t=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function jt(e,t){return(jt=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var St=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Ot(this,_t(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&jt(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){}},{key:"index",value:function(){(new it).render()}},{key:"show",value:function(){var e=new vt({id:this.params.id});this.setView("show",e),e.render(),We.all({roomId:this.params.id}).then(function(t){return e.renderMembers(t.resources)})}}])&&gt(n.prototype,r),i&&gt(n,i),t}();function Et(e){return(Et="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Pt(e,t){return!t||"object"!==Et(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function kt(e){return(kt=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function At(e,t){return(At=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}Object.assign(o.Controllers.Base.prototype,p);var It=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Pt(this,kt(t).apply(this,arguments))}return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&At(e,t)}(t,o["Controllers"].Base),t}();Object.assign(It,{Articles:Ne,Rooms:St});var Ct=It;function Tt(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}var xt=function(){function e(){!function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,e)}var t,n,r;return t=e,(n=[{key:"receivedSignal",value:function(e){switch(e.signal){case"ping":this._pingSignal();break;case"message":this._messageSignal(e)}}},{key:"_pingSignal",value:function(){o.Env.namespaceController.constructor===Ct&&alert("Ping!")}},{key:"_messageSignal",value:function(e){var t=this._getRoomView();t&&t.receivedMessage(e.message,e.author)}},{key:"_getRoomView",value:function(){return o.Env.namespaceController.constructor===Ct&&(o.Env.controller.constructor===St&&("show"===o.Env.action&&o.Env.controller.getView("show")))}}])&&Tt(t.prototype,n),r&&Tt(t,r),e}();function Rt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function Bt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}var zt=Object(T.a)({articles:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case h:return[].concat(Rt(t.payload.articles),Rt(e));case w:return Rt(t.payload.articles);case g:return[].concat(Rt(e.slice(0,t.payload.index)),[t.payload.article],Rt(e.slice(t.payload.index+1)));default:return e}},users:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case"SET_USERS":return Bt(t.payload.users);case E:return[].concat(Bt(t.payload.users),Bt(e));default:return e}}}),Dt=Object(T.b)(zt,{users:[],articles:[]});function Nt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function Lt(e){return function(e){if(Array.isArray(e)){for(var t=0,n=new Array(e.length);t<e.length;t++)n[t]=e[t];return n}}(e)||function(e){if(Symbol.iterator in Object(e)||"[object Arguments]"===Object.prototype.toString.call(e))return Array.from(e)}(e)||function(){throw new TypeError("Invalid attempt to spread non-iterable instance")}()}function qt(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){Mt(e,t,n[t])})}return e}function Mt(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}var Vt=Object(T.a)({articles:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:[],t=arguments.length>1?arguments[1]:void 0;switch(t.type){case b:return[].concat(Nt(e),Nt(t.payload.articles));case w:return Nt(t.payload.articles);case g:return[].concat(Nt(e.slice(0,t.payload.index)),[t.payload.article],Nt(e.slice(t.payload.index+1)));default:return e}},comments:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=arguments.length>1?arguments[1]:void 0;switch(t.type){case O:return qt({},e,Mt({},t.payload.articleId,[].concat(Lt(e[t.payload.articleId]),Lt(t.payload.comments))));case _:return null==e[t.payload.articleId]?e:qt({},e,Mt({},t.payload.articleId,e[t.payload.articleId].filter(function(e){return e.id!==t.payload.id})));case j:return Mt({},t.payload.articleId,t.payload.comments);case S:var n=t.payload.articleId,r=t.payload.index;return qt({},e,Mt({},n,[].concat(Lt(e[n].slice(0,r)),[t.payload.comment],Lt(e[n].slice(r+1)))));default:return e}}}),Ut=Object(T.b)(Vt,{articles:[],comments:{}}),Ft=function(e){return function(t,n){var r=arguments.length>2&&void 0!==arguments[2]?arguments[2]:{},o=t[e];r.parentId&&(o=o[r.parentId]);var i=o.find(function(e){return e.id===n});return i?[i,o.indexOf(i)]:[null,null]}},Wt=Ft("articles"),Ht=Ft("comments");function Jt(e){var t=e.comment,n=new o.Services.Date(t.createdAt).strftime("%d %b %y");return m.a.createElement("p",{id:"comment_".concat(t.id)},m.a.createElement("b",null,t.author)," on ",m.a.createElement("i",null,n)," (",m.a.createElement("a",{href:"/admin/articles/".concat(t.articleId,"/comments/").concat(t.id,"/edit")},"edit"),")",m.a.createElement("br",null),t.text)}Jt.propTypes={comment:re.a.instanceOf(te).isRequired};var Yt=Jt;function $t(e){var t=e.comments;return 0===t.length?m.a.createElement("p",null,"No comments"):m.a.createElement(m.a.Fragment,null,t.map(function(e){return m.a.createElement(Yt,{comment:e,key:e.id})}))}$t.propTypes={comments:re.a.arrayOf(re.a.instanceOf(te)).isRequired};var Gt=$t;function Kt(e){return(Kt="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Qt(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Xt(e,t){return!t||"object"!==Kt(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Zt(e){return(Zt=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function en(e,t){return(en=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var tn=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=Xt(this,Zt(t).call(this,n))).article=null,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&en(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"receivedSignal",value:function(e){var t=this;switch(e){case"updated":this.article.reload().then(function(){t.article.applyChanges(),t._renderArticle()})}}},{key:"render",value:function(e){this.article=e,this.connectWith(e),this._renderArticle()}},{key:"renderComments",value:function(e){Object(d.render)(m.a.createElement(Gt,{comments:e}),document.getElementById("comments"))}},{key:"_renderArticle",value:function(){document.getElementById("article_author").textContent=this.article.author,document.getElementById("article_title").textContent=this.article.title,document.getElementById("article_text").textContent=this.article.content}}])&&Qt(n.prototype,r),i&&Qt(n,i),t}();function nn(e){return(nn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function rn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function on(e,t){return!t||"object"!==nn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function an(e){return(an=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function cn(e,t){return(cn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var un=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),on(this,an(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&cn(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(e){e.setDefaultValuesForAdminReview(),new o.UI.Form({id:"edit_article_form",for:e}).render()}}])&&rn(n.prototype,r),i&&rn(n,i),t}();function ln(e){var t=e.article,n=new o.Services.Date(t.publishedAt).toString("short");return m.a.createElement("article",{id:"article_".concat(t.id)},m.a.createElement("h2",null,t.title),m.a.createElement("p",null,m.a.createElement("i",null,t.author," wrote this on ",n," /"," ",m.a.createElement("span",{className:"comments_quantity"},t.commentsCount," comment",1===t.commentsCount?"":"s"))),m.a.createElement("p",null,t.content),m.a.createElement("p",null,m.a.createElement("a",{href:"/admin/articles/".concat(t.id,"/edit")},"Review")))}ln.propTypes={article:re.a.instanceOf(Pe).isRequired};var fn=ln;function sn(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function pn(e){var t=sn(Object(y.useState)(e.articles),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=Dt.subscribe(function(){return r(Dt.getState().articles)});return function(){e()}});var o=n.map(function(e){return m.a.createElement(fn,{key:e.id,article:e})});return m.a.createElement(m.a.Fragment,null,o)}pn.propTypes={articles:re.a.arrayOf(re.a.instanceOf(Pe)).isRequired};var yn=pn;function mn(e){return(mn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function dn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function bn(e,t){return!t||"object"!==mn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function hn(e){return(hn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function vn(e,t){return(vn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var wn=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),bn(this,hn(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&vn(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"published",value:function(){Pe.get("published").then(function(e){Dt.dispatch(k(e.resources)),Object(d.render)(m.a.createElement(yn,{articles:e.resources}),document.getElementById("articles"))})}},{key:"edit",value:function(){var e=new tn;Pe.find(this.params.id).then(function(t){e.render(t),(new un).render(t)}),te.all({articleId:this.params.id}).then(function(t){e.renderComments(t.resources)})}}])&&dn(n.prototype,r),i&&dn(n,i),t}();function gn(e){return(gn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function On(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function _n(e,t){return!t||"object"!==gn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function jn(e){return(jn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Sn(e,t){return(Sn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var En=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),_n(this,jn(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Sn(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{},t=new o.UI.Form({for:new te({id:e.commentId,resource:"admin"}),id:"edit_comment_".concat(e.commentId),initObj:!0});t.render(),o.Env.test={commentFormObj:t.getObj()}}}])&&On(n.prototype,r),i&&On(n,i),t}();function Pn(e){return(Pn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function kn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function An(e,t){return!t||"object"!==Pn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function In(e){return(In=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Cn(e,t){return(Cn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Tn=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),An(this,In(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Cn(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"edit",value:function(){(new En).render({commentId:this.params.id})}}])&&kn(n.prototype,r),i&&kn(n,i),t}();function xn(e){return(xn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Rn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Bn(e,t){return!t||"object"!==xn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function zn(e){return(zn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Dn(e,t){return(Dn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Nn=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Bn(this,zn(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Dn(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({id:"sign_in_admin",delegator:this,callbackSuccess:"_signedIn"}).render()}},{key:"_signedIn",value:function(){window.location.href="/admin"}}])&&Rn(n.prototype,r),i&&Rn(n,i),t}();function Ln(e){return(Ln="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function qn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Mn(e,t){return!t||"object"!==Ln(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Vn(e){return(Vn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Un(e,t){return(Un=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Fn=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Mn(this,Vn(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Un(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"new",value:function(){(new Nn).render()}}])&&qn(n.prototype,r),i&&qn(n,i),t}(),Wn=function(e){var t=e.user;return m.a.createElement("tr",{id:"user_".concat(t.id)},m.a.createElement("td",null,t.email),m.a.createElement("td",null,t.username),m.a.createElement("td",{className:"confirmed"},t.confirmed?"Yes":"No"),m.a.createElement("td",null,m.a.createElement("a",{href:"/admin/users/".concat(t.id)},"Show")," |"," ",m.a.createElement("a",{href:"/admin/users/".concat(t.id,"/edit")},"Edit")," |"," ",m.a.createElement("a",{href:"/admin/users/".concat(t.id),"data-method":"delete","data-confirm":"Are you sure?"},"Delete")," ","|"," ",m.a.createElement("a",{href:"#",onClick:function(e){return function(e,t){e.preventDefault(),o.Env.loco.emit({signal:"ping",user_id:t})}(e,t.id)}},"Ping")))};Wn.propTypes={user:re.a.instanceOf(Ze).isRequired};var Hn=Wn;function Jn(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Yn(e){var t=Jn(Object(y.useState)(e.users),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=Dt.subscribe(function(){return r(Dt.getState().users)});return function(){e()}});var o=n.map(function(e){return m.a.createElement(Hn,{key:"user_".concat(e.id),user:e})});return m.a.createElement(m.a.Fragment,null,o)}Yn.propTypes={users:re.a.arrayOf(re.a.instanceOf(Ze)).isRequired};var $n=Yn;function Gn(e){return(Gn="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Kn(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Qn(e,t){return!t||"object"!==Gn(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Xn(e){return(Xn=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Zn(e,t){return(Zn=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var er=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=Qn(this,Xn(t).call(this,n))).user=n.user,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Zn(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){document.getElementById("user_email").textContent=this.user.email,document.getElementById("user_username").textContent=this.user.username,document.getElementById("user_confirmed").textContent=this.user.confirmed?"Yes":"No",this._updateEditLink()}},{key:"_updateEditLink",value:function(){var e=document.getElementById("edit_link"),t=e.getAttribute("href");e.setAttribute("href",t.replace("/0/","/#{@user.id}/"))}}])&&Kn(n.prototype,r),i&&Kn(n,i),t}();function tr(e){return(tr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function nr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function rr(e,t){return!t||"object"!==tr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function or(e){return(or=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function ir(e,t){return(ir=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var ar=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=rr(this,or(t).call(this,n))).user=n.user,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&ir(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({for:this.user,initObj:!0,id:"admin_user_form"}).render()}}])&&nr(n.prototype,r),i&&nr(n,i),t}();function cr(e){return(cr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function ur(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function lr(e,t){return!t||"object"!==cr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function fr(e){return(fr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function sr(e,t){return(sr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var pr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),lr(this,fr(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&sr(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"index",value:function(){Ze.get("all").then(function(e){var t;Dt.dispatch((t=e.resources,{type:E,payload:{users:t}})),Object(d.render)(m.a.createElement($n,{users:e.resources}),document.querySelector("table tbody"))})}},{key:"show",value:function(){Ze.find(this.params.id).then(function(e){return new er({user:e}).render()})}},{key:"edit",value:function(){new ar({user:new Ze({id:this.params.id})}).render()}}])&&ur(n.prototype,r),i&&ur(n,i),t}();function yr(e){return(yr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function mr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function dr(e,t){return!t||"object"!==yr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function br(e){return(br=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function hr(e,t){return(hr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}Object.assign(o.Controllers.Base.prototype,p);var vr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),dr(this,br(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&hr(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){this.setScope("admin")}}])&&mr(n.prototype,r),i&&mr(n,i),t}();Object.assign(vr,{Articles:wn,Comments:Tn,Sessions:Fn,Users:pr});var wr=vr;function gr(e){var t=e.comment,n=new o.Services.Date(t.createdAt).strftime("%d %b %y");return m.a.createElement("p",{id:"comment_".concat(t.id)},m.a.createElement("b",null,t.author)," on ",m.a.createElement("i",null,n)," wrote:",m.a.createElement("br",null),t.text)}gr.propTypes={comment:re.a.instanceOf(te).isRequired};var Or=gr;function _r(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function jr(e){var t=e.articleId,n=_r(Object(y.useState)(e.comments),2),r=n[0],o=n[1];Object(y.useEffect)(function(){var e=Ut.subscribe(function(){o(oe(Ut.getState(),t))});return function(){e()}});var i=r.map(function(e){return m.a.createElement(Or,{key:e.id,comment:e})});return 0===r.length?m.a.createElement("p",{id:"no_comments"},"No comments."):m.a.createElement(m.a.Fragment,null,i)}jr.propTypes={articleId:re.a.number.isRequired,comments:re.a.arrayOf(re.a.instanceOf(te)).isRequired};var Sr=jr;function Er(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Pr(e){var t=Er(Object(y.useState)(e.comments),2),n=t[0],r=t[1];return Object(y.useEffect)(function(){var t=Ut.subscribe(function(){r(oe(Ut.getState(),e.articleId))});return function(){t()}}),m.a.createElement(m.a.Fragment,null,"".concat(n.length," comment").concat(1===n.length?"":"s"))}Pr.propTypes={articleId:re.a.number.isRequired,comments:re.a.arrayOf(re.a.instanceOf(te)).isRequired};var kr=Pr;function Ar(e){return(Ar="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Ir(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Cr(e,t){return!t||"object"!==Ar(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Tr(e){return(Tr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function xr(e,t){return(xr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Rr=function(e){function t(){var e,n=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),(e=Cr(this,Tr(t).call(this,n))).article=null,e.newComment=n.comment,e}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&xr(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({for:this.newComment,initObj:!0,id:"new_comment"}).render()}},{key:"renderArticle",value:function(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:null;this.article||(this.article=e,this.connectWith(this.article,{receiver:"_articleReceivedSignal"})),document.getElementById("title").textContent=this.article.title,document.getElementById("author").textContent=this.article.author;var t=new o.Services.Date(this.article.publishedAt);document.getElementById("pub_date").textContent=t.toString("short");var n=document.getElementById("text");n.innerHTML="";var r=(new o.Helpers.Text).simpleFormat(this.article.content);n.insertAdjacentHTML("beforeend",r)}},{key:"_articleReceivedSignal",value:function(e){var t=this;switch(e){case"updating":new s({warning:"Author is currently editing article. Be aware of possible changes."}).render();break;case"updated":this.article.reload().then(function(){t.article.applyChanges(),t.renderArticle()})}}}])&&Ir(n.prototype,r),i&&Ir(n,i),t}();function Br(e){return(Br="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function zr(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Dr(e,t){return!t||"object"!==Br(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Nr(e){return(Nr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Lr(e,t){return(Lr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var qr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Dr(this,Nr(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Lr(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"show",value:function(){var e=this,t=new te({articleId:this.params.id}),n=new Rr({comment:t});n.render(),Pe.find(this.params.id).then(function(e){Ut.dispatch(k([e])),n.renderArticle(e)}),te.get("count",{articleId:this.params.id}).then(function(t){te.all({articleId:e.params.id,total:t.total}).then(function(t){Ut.dispatch(I(t,e.params.id)),Object(d.render)(m.a.createElement(Sr,{articleId:e.params.id,comments:t}),document.getElementById("comments")),Object(d.render)(m.a.createElement(kr,{articleId:e.params.id,comments:t}),document.getElementById("comments_count"))})})}}])&&zr(n.prototype,r),i&&zr(n,i),t}();function Mr(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Vr(){var e=Mr(Object(y.useState)(1),2),t=e[0],n=e[1],r=Mr(Object(y.useState)(!1),2),o=r[0],i=r[1];function a(e){e.preventDefault();var r=t+1;n(r),Pe.get("all",{page:r}).then(function(e){e.resources.length>0?Ut.dispatch(P(e.resources)):i(!0)}).catch(function(e){return alert("Invalid URL: ".concat(e))})}return o?m.a.createElement("span",null,"No more posts."):m.a.createElement("a",{href:"#",id:"load_more",onClick:a},"Load more…")}function Ur(e){var t=e.article,n=new o.Services.Date(t.publishedAt).toString("short");return m.a.createElement("article",{id:"article_".concat(t.id)},m.a.createElement("h2",null,t.title),m.a.createElement("p",null,m.a.createElement("i",null,t.author," wrote this on ",n," /"," ",m.a.createElement("a",{href:"/articles/".concat(t.id,"#comments"),className:"comments_quantity"},t.commentsCount," comment",1===t.commentsCount?"":"s"))),m.a.createElement("p",null,t.content),m.a.createElement("p",null,m.a.createElement("a",{href:"/articles/".concat(t.id)},"Continued…")))}Ur.propTypes={article:re.a.instanceOf(Pe).isRequired};var Fr=Ur;function Wr(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}function Hr(e){var t=Wr(Object(y.useState)(e.articles),2),n=t[0],r=t[1];Object(y.useEffect)(function(){var e=Ut.subscribe(function(){return r(Ut.getState().articles)});return function(){e()}});var o=n.map(function(e){return m.a.createElement(Fr,{key:e.id,article:e})});return m.a.createElement(m.a.Fragment,null,o)}Hr.propTypes={articles:re.a.arrayOf(re.a.instanceOf(Pe)).isRequired};var Jr=Hr;function Yr(e){return(Yr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function $r(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Gr(e,t){return!t||"object"!==Yr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function Kr(e){return(Kr=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Qr(e,t){return(Qr=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var Xr=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Gr(this,Kr(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Qr(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"index",value:function(){Object(d.render)(m.a.createElement(Vr,null),document.getElementById("load_more_wrapper")),Pe.get("all",{page:1}).then(function(e){Ut.dispatch(k(e.resources)),Object(d.render)(m.a.createElement(Jr,{articles:e.resources}),document.getElementById("articles"))})}}])&&$r(n.prototype,r),i&&$r(n,i),t}();function Zr(e){return(Zr="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function eo(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function to(e,t){return!t||"object"!==Zr(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function no(e){return(no=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function ro(e,t){return(ro=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var oo=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),to(this,no(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&ro(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"render",value:function(){new o.UI.Form({for:new Ze,delegator:this,callbackSuccess:"_created"}).render()}},{key:"receivedSignal",value:function(e){switch(e){case"confirming":this._confirming();break;case"confirmed":this._confirmed()}}},{key:"_created",value:function(e){this.connectWith(new Ze({id:e.id})),document.querySelector("form").style.display="none",document.getElementById("sign_in_paragraph").classList.remove("none"),document.getElementById("verification_info").classList.remove("none"),new s({notice:e.notice}).render()}},{key:"_confirming",value:function(){document.getElementById("verification_info").textContent=document.getElementById("verification_progress").textContent}},{key:"_confirmed",value:function(){window.location.href="/user/sessions/new?event=confirmed"}}])&&eo(n.prototype,r),i&&eo(n,i),t}();function io(e){return(io="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function ao(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function co(e,t){return!t||"object"!==io(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function uo(e){return(uo=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function lo(e,t){return(lo=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var fo=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),co(this,uo(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&lo(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"new",value:function(){(new oo).render()}}])&&ao(n.prototype,r),i&&ao(n,i),t}();function so(e){return(so="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function po(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function yo(e,t){return!t||"object"!==so(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function mo(e){return(mo=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function bo(e,t){return(bo=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}Object.assign(o.Controllers.Base.prototype,p);var ho=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),yo(this,mo(t).apply(this,arguments))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&bo(e,t)}(t,o["Controllers"].Base),n=t,(r=[{key:"initialize",value:function(){this.setScope("main")}}])&&po(n.prototype,r),i&&po(n,i),t}();Object.assign(ho,{Articles:qr,Pages:Xr,Users:fo});var vo=ho;function wo(e){return(wo="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function go(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Oo(e,t){return!t||"object"!==wo(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function _o(e){return(_o=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function jo(e,t){return(jo=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}function So(e,t,n){return t in e?Object.defineProperty(e,t,{value:n,enumerable:!0,configurable:!0,writable:!0}):e[t]=n,e}function Eo(e,t){return function(e){if(Array.isArray(e))return e}(e)||function(e,t){var n=[],r=!0,o=!1,i=void 0;try{for(var a,c=e[Symbol.iterator]();!(r=(a=c.next()).done)&&(n.push(a.value),!t||n.length!==t);r=!0);}catch(e){o=!0,i=e}finally{try{r||null==c.return||c.return()}finally{if(o)throw i}}return n}(e,t)||function(){throw new TypeError("Invalid attempt to destructure non-iterable instance")}()}var Po=function(e){var t=e.id;o.Env.namespaceController.constructor===wr?Pe.find({id:t,abbr:!0,resource:"admin"}).then(function(e){Dt.dispatch({type:h,payload:{articles:[e]}})}):Pe.find({id:t,abbr:!0}).then(function(e){return Ut.dispatch(P([e]))})},ko=function(e,t){var n=e.article_id,r=Ut;o.Env.namespaceController.constructor===wr?r=Dt:o.Env.namespaceController.constructor===Ct&&(r=N);var i=Eo(Wt(r.getState(),n),2),a=i[0],c=i[1];if(a){var u=new Pe(function(e){for(var t=1;t<arguments.length;t++){var n=null!=arguments[t]?arguments[t]:{},r=Object.keys(n);"function"==typeof Object.getOwnPropertySymbols&&(r=r.concat(Object.getOwnPropertySymbols(n).filter(function(e){return Object.getOwnPropertyDescriptor(n,e).enumerable}))),r.forEach(function(t){So(e,t,n[t])})}return e}({},a,{commentsCount:a.commentsCount+t}));r.dispatch(A(u,c))}},Ao=function(e){var t=e.article_id,n=e.id,r=N,i={articleId:t,id:n};o.Env.namespaceController.constructor===vo&&(r=Ut,i.resource="main"),Eo(Wt(r.getState(),t),1)[0]&&te.find(i).then(function(e){return r.dispatch(function(e,t){return{type:O,payload:{articleId:t,comments:e}}}([e],t))})},Io=function(e){var t=e.article_id,n=e.id,r=Ut;o.Env.namespaceController.constructor===Ct&&(r=N),r.dispatch(function(e,t){return{type:_,payload:{id:e,articleId:t}}}(n,t))},Co=function(e){function t(){var e=arguments.length>0&&void 0!==arguments[0]?arguments[0]:{};return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Oo(this,_o(t).call(this,e))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&jo(e,t)}(t,o["Views"].Base),n=t,(r=[{key:"receivedSignal",value:function(e,t){switch(e){case"Article created":n=t.id,o.Env.namespaceController.constructor===Ct&&Pe.find({id:n,abbr:!0}).then(function(e){return N.dispatch(P([e]))});break;case"Article published":Po(t);break;case"Article updated":!function(e){var t=e.id,n={id:t,abbr:!0},r=Ut;o.Env.namespaceController.constructor===wr?(n.resource="admin",r=Dt):o.Env.namespaceController.constructor===Ct&&(r=N);var i=Eo(Wt(r.getState(),t),2),a=i[0],c=i[1];a&&Pe.find(n).then(function(e){return r.dispatch(A(e,c))})}(t);break;case"Article.Comment created":ko(t,1),Ao(t);break;case"Article.Comment destroyed":ko(t,-1),Io(t);break;case"Article.Comment updated":!function(e){var t=e.article_id,n=e.id,r=Ut;o.Env.namespaceController.constructor===Ct&&(r=N);var i=Eo(Ht(r.getState(),n,{parentId:t}),2),a=i[0],c=i[1];a&&a.reload().then(function(){a.applyChanges(),r.dispatch(C(a,t,c))})}(t);break;case"User created":Ze.find(t.id).then(function(e){return Dt.dispatch({type:E,payload:{users:[e]}})})}var n}},{key:"call",value:function(){this.connectWith([Pe,te,Ze])}}])&&go(n.prototype,r),i&&go(n,i),t}();o.Deps.cable=Object(i.createConsumer)(),o.Deps.NotificationCenter=xt;var To=new Co;new o.Loco({notifications:{enable:!0,log:!0,size:10},postInit:function(){To.call(),"test"===document.querySelector("body").getAttribute("data-rails-env")&&o.Env.loco.getWire().setPollingTime(1e3)}}).init();n(138);function xo(e){return(xo="function"==typeof Symbol&&"symbol"==typeof Symbol.iterator?function(e){return typeof e}:function(e){return e&&"function"==typeof Symbol&&e.constructor===Symbol&&e!==Symbol.prototype?"symbol":typeof e})(e)}function Ro(e,t){for(var n=0;n<t.length;n++){var r=t[n];r.enumerable=r.enumerable||!1,r.configurable=!0,"value"in r&&(r.writable=!0),Object.defineProperty(e,r.key,r)}}function Bo(e,t){return!t||"object"!==xo(t)&&"function"!=typeof t?function(e){if(void 0===e)throw new ReferenceError("this hasn't been initialised - super() hasn't been called");return e}(e):t}function zo(e){return(zo=Object.setPrototypeOf?Object.getPrototypeOf:function(e){return e.__proto__||Object.getPrototypeOf(e)})(e)}function Do(e,t){return(Do=Object.setPrototypeOf||function(e,t){return e.__proto__=t,e})(e,t)}var No,Lo,qo,Mo=function(e){function t(){return function(e,t){if(!(e instanceof t))throw new TypeError("Cannot call a class as a function")}(this,t),Bo(this,zo(t).call(this))}var n,r,i;return function(e,t){if("function"!=typeof t&&null!==t)throw new TypeError("Super expression must either be null or a function");e.prototype=Object.create(t&&t.prototype,{constructor:{value:e,writable:!0,configurable:!0}}),t&&Do(e,t)}(t,o["Validators"].Base),n=t,(r=[{key:"validate",value:function(){if(null!=this.val)switch(xo(this.val)){case"string":var e=this._getVulgarWord();if(!new RegExp("".concat(e),["i"]).exec(this.val))return;this._addErrorMessage();break;default:throw new TypeError("Vulgarity validator is applicable only for strings and ".concat(this.attr," isn't."))}}},{key:"_addErrorMessage",value:function(){var e=o.I18n[o.Env.loco.getLocale()].errors.messages.vulgarity;this.obj.addErrorMessage(e,{for:this.attr})}},{key:"_getVulgarWord",value:function(){switch(o.Env.loco.getLocale()){case"pl":return"kurwa";case"en":return"fuck"}}}])&&Ro(n.prototype,r),i&&Ro(n,i),t}();qo="Vulgarity",(Lo="identity")in(No=Mo)?Object.defineProperty(No,Lo,{value:qo,enumerable:!0,configurable:!0,writable:!0}):No[Lo]=qo;var Vo=Mo;Object.assign(o.Validators,{Vulgarity:Vo}),o.I18n.pl={variants:{few:function(e){var t=parseInt(o.Utils.String.last(String(e)));-1!==[2,3,4].indexOf(t)&&(2!==String(e).length||String(e)[0])}},models:{},attributes:{},ui:{form:{sending:"Wysyłam...",success:"Sukces",errors:{connection:"Błąd z połączeniem",invalid_data:"Nieprawidłowe dane"}}},date:{formats:{default:"%d-%m-%Y",short:"%d %b",long:"%B %d, %Y"},day_names:["niedziela","poniedziałek","wtorek","środa","czwartek","piątek","sobota"],abbr_day_names:["nie","pon","wto","śro","czw","pią","sob"],month_names:["styczeń","luty","marzec","kwiecień","maj","czerwiec","lipiec","sierpień","wrzesień","październik","listopad","grudzień"],abbr_month_names:["sty","lut","mar","kwi","maj","cze","lip","sie","wrz","paź","lis","gru"]},errors:{messages:{accepted:"musi zostać zaakceptowane",blank:"nie może być puste",confirmation:"nie zgadza się z polem %{attribute}",empty:"nie może być puste",equal_to:"musi być równe %{count}",even:"musi być parzyste",exclusion:"jest zarezerwowane",greater_than:"musi być większe od %{count}",greater_than_or_equal_to:"musi być większe lub równe %{count}",inclusion:"nie znajduje się na liście dopuszczalnych wartości",invalid:"jest nieprawidłowe",less_than:"musi być mniejsze od %{count}",less_than_or_equal_to:"musi być mniejsze lub równe %{count}",not_a_number:"nie jest liczbą",not_an_integer:"musi być liczbą całkowitą",odd:"musi być nieparzyste",present:"musi być puste",too_long:{few:"jest za długie (maksymalnie %{count} znaki)",many:"jest za długie (maksymalnie %{count} znaków)",one:"jest za długie (maksymalnie jeden znak)",other:"jest za długie (maksymalnie %{count} znaków)"},too_short:{few:"jest za krótkie (przynajmniej %{count} znaki)",many:"jest za krótkie (przynajmniej %{count} znaków)",one:"jest za krótkie (przynajmniej jeden znak)",other:"jest za krótkie (przynajmniej %{count} znaków)"},wrong_length:{few:"ma nieprawidłową długość (powinna wynosić %{count} znaki)",many:"ma nieprawidłową długość (powinna wynosić %{count} znaków)",one:"ma nieprawidłową długość (powinna wynosić jeden znak)",other:"ma nieprawidłową długość (powinna wynosić %{count} znaków)"},other_than:"musi być inna niż %{count}"}}},o.I18n.pl.models.User="Użytkownik",o.I18n.pl.attributes.User={password:"Hasło",passwordConfirmation:"Powtórz hasło"},o.I18n.en.errors.messages.vulgarity="contains strong language.",o.I18n.pl.errors.messages.vulgarity="zawiera mocny język.",Object.assign(o.Controllers,{Admin:wr,Main:vo,User:Ct}),Object.assign(Pe,{Comment:te}),Object.assign(pt,{Member:We}),Object.assign(o.Models,{Article:Pe,Room:pt,User:Ze}),window.test={Channels:o.Channels,Env:o.Env}}});
 //# sourceMappingURL=application.js.map
