--- test/dummy/frontend/js/actions/user.js
@@ -1,13 +1,0 @@
-/*
- * action types
- */
-
-export const REMOVE_ARTICLE = "REMOVE_ARTICLE";
-
-/*
- * action creators
- */
-
-export function removeArticle(id) {
-  return { type: REMOVE_ARTICLE, payload: { id } };
-}
