+++ test/dummy/frontend/js/components/user/Article.js
@@ -3,7 +3,7 @@ import PropTypes from "prop-types";
 
 import ArticleModel from "models/Article";
 
-import { removeArticle } from "actions/user";
+import { removeArticle } from "actions";
 import store from "stores/user";
 
 function Article({ article, onArticleDestroyed }) {
