+++ test/dummy/frontend/js/reducers/main/comments.js
@@ -1,4 +1,9 @@
-import { ADD_COMMENTS, REMOVE_COMMENT, SET_COMMENTS } from "actions/shared";
+import {
+  ADD_COMMENTS,
+  REMOVE_COMMENT,
+  SET_COMMENTS,
+  UPDATE_COMMENT
+} from "actions";
 
 export default (state = {}, action) => {
   switch (action.type) {
@@ -20,14 +25,15 @@ export default (state = {}, action) => {
       };
     case SET_COMMENTS:
       return { [action.payload.articleId]: action.payload.comments };
-    case "UPDATE_COMMENT": {
+    case UPDATE_COMMENT: {
       const articleId = action.payload.articleId;
+      const index = action.payload.index;
       return {
         ...state,
         [articleId]: [
-          ...state[articleId].slice(0, action.payload.index),
+          ...state[articleId].slice(0, index),
           action.payload.comment,
-          ...state[articleId].slice(action.payload.index + 1)
+          ...state[articleId].slice(index + 1)
         ]
       };
     }
